### 2024-04-02 #6568 @editorialbot
[PRE REVIEW]: GeneScape: A Python package for gene ontology analysis
**Submitting author:** @ialbert (<a href="http://orcid.org/0000-0001-8366-984X">Istvan Albert</a>)
**Repository:** https://github.com/ialbert/genescape-central
**Branch with paper.md** (empty if default branch):
**Version:** 0.7.2
