### 2024-04-02 #6568 @Kevin-Mattheus-Moerman
@ialbert, thanks for this submission. I am the AEiC on this track and here to help process the initial steps. Before we proceed, please can you have a look at the following points:

- [ ] It would also be good to link to testing and documentation from the README. Testing (preferably automated) is a requirement for JOSS.
- [ ] I think perhaps the above paper compilation error is due to the way you have this: `affiliation: ["1", "2"] `, I think you should remove the square brackets and have it quoted like this: `affiliation: "1, 2"` i.e. as per https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography. You can retry the paper compilation by calling `@editorialbot generate pdf`
