### 2024-04-02 85335e2
updates to the paper, JOSS checklist
