+++ docs/paper.bib
@@ -62,6 +62,7 @@
   number={1},
   pages={1--17},
   year={2018},
+  doi={10.1038/s41598-018-28948-z},
   publisher={Nature Publishing Group}
 }
 
