+++ CONTRIBUTING.md
@@ -1,0 +1,13 @@
+# Contributing to GeneScape
+
+> Be well, do good work, and keep in touch.
+> 
+> -- Garrison Keillor
+
+Whether you're fixing a bug as quietly as snow falling on cedars, or adding a feature as boldly as a prairie sunrise, remember: 
+
+1. be well, for kindness nurtures both code and coder; 
+2. do good work, for craftsmanship echoes in eternity; 
+3. and keep in touch, for in the sharing of updates, issues, and pull requests, we find the strength of our community. 
+
+As the seasons roll from spring commits to autumn releases, let us walk this path together, creating software that, like a good story, enriches all who encounter it.
