### 2024-04-02 28a2bdc
more updates to the paper, JOSS checklist
