+++ docs/paper.bib
@@ -76,4 +76,19 @@
   url = {https://bioconductor.org/packages/topGO}
 }
 
+@article{QuickGO,
+    author = {Binns, David and Dimmer, Emily and Huntley, Rachael and Barrell, Daniel and O'Donovan, Claire and Apweiler, Rolf},
+    title = "{QuickGO: a web-based tool for Gene Ontology searching}",
+    journal = {Bioinformatics},
+    volume = {25},
+    number = {22},
+    pages = {3045-3046},
+    year = {2009},
+    month = {09},
+    abstract = "{Summary: QuickGO is a web-based tool that allows easy browsing of the Gene Ontology (GO) and all associated electronic and manual GO annotations provided by the GO Consortium annotation groups QuickGO has been a popular GO browser for many years, but after a recent redevelopment it is now able to offer a greater range of facilities including bulk downloads of GO annotation data which can be extensively filtered by a range of different parameters and GO slim set generation.Availability and Implementation: QuickGO has implemented in JavaScript, Ajax and HTML, with all major browsers supported. It can be queried online at http://www.ebi.ac.uk/QuickGO. The software for QuickGO is freely available under the Apache 2 licence and can be downloaded from http://www.ebi.ac.uk/QuickGO/installation.htmlContact:  goa@ebi.ac.uk; dbinns@ebi.ac.uk}",
+    issn = {1367-4803},
+    doi = {10.1093/bioinformatics/btp536},
+    url = {https://doi.org/10.1093/bioinformatics/btp536},
+    eprint = {https://academic.oup.com/bioinformatics/article-pdf/25/22/3045/48997998/bioinformatics\_25\_22\_3045.pdf},
+}
 
