### 2024-04-02 7cbe60e
more updates to the paper, JOSS checklist
