### 2024-04-02 ca6f145
more updates to the paper, JOSS checklist
