### 2024-04-02 20ac775
more updates to the paper, JOSS checklist
