### 2024-04-02 1b4895e
image resize, JOSS format
