### 2024-04-02 dc3eda4
image resize, JOSS format
