### 2024-04-02 8cb0dff
image resize, JOSS format
