### 2024-04-02 #6568 @ialbert
@Kevin-Mattheus-Moerman I done my best to address the issues

> * [x]  Please study the above reference check ☝️ and see if you can address any of the reported potential DOI issues. You can add/amend DOI entries in your `.bib` file, and call `@editorialbot check references` here to check them again.

The missing DOI entry has been added, the reference check now passes.

> * [x]  Your project appears to lack contributing guidelines at the moment. Please work to add these and to mention/link to them in your README. You could for instance create a `CONTRIBUTING.md` file (see here for some examples: https://contributing.md/example/).

`CONTRIBUTING.md` has been added and is cited in the README.md

> * [x]  It would also be good to link to testing and documentation from the README. Testing (preferably automated) is a requirement for JOSS.

Testing has always been part of the software via a `Makefile` and this information is now added to the `README.md`.

Completely automated testing in the traditional software engineering approach is less feasible because the graph drawing process is not fully deterministic. The same input may result in graphs with the same topology but different visual layouts. In addition, testing GUI interfaces is far more complex task than traditional testing of inputs and outs.

I will be exploring better testing approaches in subsequent releases;  there are intermediate steps that can be tested.

> * [x]  I think perhaps the above paper compilation error is due to the way you have this: `affiliation: ["1", "2"] `, I think you should remove the square brackets and have it quoted like this: `affiliation: "1, 2"` i.e. as per https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography. You can retry the paper compilation by calling `@editorialbot generate pdf`

The references have been corrected.
