### 2024-04-02 0b7d212
adding automated testing
