### 2024-04-02 ec206f0
the graphs are stable, the system generates sorted node orders
