+++ src/genescape/server.py
@@ -1,16 +1,22 @@
-import sys, os, functools, webbrowser, threading
+import functools
+import os
+import pprint
+import sys
+import threading
+import time
+import webbrowser
 from random import random
-import time, toml, pprint
-from genescape import tree, annot, utils, resources
-from genescape.bottle import Bottle, static_file, template, view
-from genescape.bottle import TEMPLATE_PATH
-from genescape import bottle
-from genescape.bottle import get, post, request, response, redirect
+
+import toml
+
 from genescape import __version__ as VERSION
+from genescape import annot, bottle, resources, tree, utils
+from genescape.bottle import TEMPLATE_PATH, Bottle, get, post, redirect, request, response, static_file, template, view
 
 DEVMODE = True
 
 CURR_DIR = os.path.dirname(os.path.realpath(__file__))
+
 
 def init_server(devmode=False):
 
@@ -28,20 +34,21 @@ def init_server(devmode=False):
     return res
 
 
-
 app = Bottle()
+
 
 def debugger(func):
     """
     Decorator that prints the execution time of the function it decorates.
     """
+
     @functools.wraps(func)
     def wrapper(*args, **kwargs):
         if DEVMODE:
 
-            #print(f"# method: {request.method}")
+            # print(f"# method: {request.method}")
 
-            #for header, value in request.headers.items():
+            # for header, value in request.headers.items():
             #    print(f"# header: {header}: {value}")
 
             for key in request.query.keys():
@@ -52,7 +59,9 @@ def debugger(func):
 
         result = func(*args, **kwargs)
         return result
+
     return wrapper
+
 
 def textarea(request, name='input'):
     text = request.forms.get(name, '')
@@ -61,16 +70,19 @@ def textarea(request, name='input'):
     terms = list(terms)
     return terms
 
+
 def get_param(request, name='param'):
     value = request.forms.get(name, '')
     value = value.strip()
     return value
+
 
 def safe_int(value, default=1):
     try:
         return int(value)
     except:
         return default
+
 
 def webapp(res, config, devmode=False, host="localhost", port=8000):
 
@@ -109,7 +121,7 @@ def webapp(res, config, devmode=False, host="localhost", port=8000):
         # Choose a different index
         index = res.get_index(db)
 
-        utils.info(f"code={db}, file={str(index)}")
+        utils.info(f"code={db}, file={index!s}")
 
         # Generate the output.
         graph, ann = tree.parse_input(inp=inp, index=index, pattern=patt, root=root, mincount=count)
@@ -125,7 +137,6 @@ def webapp(res, config, devmode=False, host="localhost", port=8000):
 
         return param
 
-
     # Setting the debug mode.
     if devmode:
         bottle.debug(True)
@@ -136,6 +147,7 @@ def webapp(res, config, devmode=False, host="localhost", port=8000):
     except Exception as e:
         print(f"Server Error: {e}", sys.stderr)
 
+
 def get_dbs(config):
     dbs = []
     for value in config.get("index", []):
@@ -144,6 +156,7 @@ def get_dbs(config):
         selected = value.get("selected", '')
         dbs.append((code, name, selected))
     return dbs
+
 
 def start(devmode=False, reset=False, index=None, host="localhost", port=8000, proto="http", config=None):
     global TEMPLATE_PATH, DEVMODE
@@ -172,6 +185,7 @@ def start(devmode=False, reset=False, index=None, host="localhost", port=8000, p
     utils.info(f"server dir: {res.WEB_DIR}")
 
     webapp(res=res, devmode=devmode, host=host, port=port, config=config)
+
 
 if __name__ == "__main__":
     start()
