### 2024-04-02 0413e84
the graphs are stable, the system generates sorted node orders
