+++ src/genescape/__about__.py
@@ -1,1 +1,1 @@
-__version__ = "0.7.2"
+__VERSION__ = "0.8.0"
