+++ pyproject.toml
@@ -162,7 +162,7 @@ ignore = [
   "C901", "PLR0911", "PLR0912", "PLR0913", "PLR0915",
 
   # Pet peevees
-  "Q000", "C402",  "F401", "E401",  "T201", "E501", 
+  "Q000", "C402",  "F401", "E401",  "T201", "E501", "S311", "RUF005", "S603",
 ]
 unfixable = [
   # Don't touch unused imports
