+++ Makefile
@@ -26,10 +26,6 @@ test:
 testall: test
 	(cd test && make test)
 
-# A full test with file generation.
-testall: test
-	(cd test && make test)
-
 # Runs a linter.
 lint:
 	hatch run lint:style
