### 2024-04-03 #6568 @mahfuz05062
@Kevin-Mattheus-Moerman Yes, I have good experience working with gene ontology data and I can edit this paper.
