+++ docs/paper.md
@@ -11,9 +11,9 @@ authors:
     affiliation: "1, 2"
 
 affiliations:
- - name: Bioinformatics Consulting Center, Pennsylvania State University, USA
+ - name: Bioinformatics Consulting Center, Pennsylvania State University, United States of America
    index: 1
- - name: Biochemistry and Molecular Biology, Pennsylvania State University, USA
+ - name: Biochemistry and Molecular Biology, Pennsylvania State University, United States of America
    index: 2
    
 
