### 2024-04-03 e513ec0
small fixes to the paper
