### 2024-04-03 e000fd0
fixes to automated testing framework
