### 2024-04-03 e5d5855
updates to testing framework
