+++ .github/workflows/release-builder.yml
@@ -15,7 +15,7 @@ permissions:
 jobs:
   build:
 
-    runs-on: windows-latest
+    runs-on: ubuntu-latest
 
     steps:
     - uses: actions/checkout@v3
@@ -23,14 +23,24 @@ jobs:
       uses: actions/setup-python@v3
       with:
         python-version: "3.10"
+
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install pytest
-        pip install . 
+        pip install pytest pyinstaller 
+        pip install .
+
+    - name: Install the package
+      run: pip install .
+
     - name: Test with pytest
-      run: |
-        pytest
-    - name: Create package
-      run: |
-        buildwin.bat
+      run: pytest
+
+    - name: Create the package
+      run: python src/genescape/dist.py
+
+    - name: Upload resulting file as an artifact
+      uses: actions/upload-artifact@v3
+      with:
+        name: genescape-package
+        path: dist/GeneScape-latest.zip
