+++ .github/workflows/release-builder.yml
@@ -15,7 +15,7 @@ permissions:
 jobs:
   build:
 
-    runs-on: ubuntu-latest
+    runs-on: windows-latest
 
     steps:
     - uses: actions/checkout@v3
@@ -42,5 +42,5 @@ jobs:
     - name: Upload resulting file as an artifact
       uses: actions/upload-artifact@v3
       with:
-        name: genescape-package
+        name: GeneScape-latest.zip
         path: dist/GeneScape-latest.zip
