### 2024-04-03 d87de03
updates artifact generation
