+++ .github/workflows/release-builder.yml
@@ -18,29 +18,25 @@ jobs:
     runs-on: windows-latest
 
     steps:
-    - uses: actions/checkout@v3
+    - uses: actions/checkout@v4
     - name: Set up Python 3.10
-      uses: actions/setup-python@v3
+      uses: actions/setup-python@v4
       with:
         python-version: "3.10"
 
-    - name: Install dependencies
+    - name: Set up the environment
       run: |
         python -m pip install --upgrade pip
         pip install pytest pyinstaller 
         pip install .
+        pytest
+        python src/genescape/dist.py
 
-    - name: Install the package
-      run: pip install .
+    - name: Extract package version
+      run: echo "VERSION=$(python -c 'from genescape import __version__; print(__version__)')" >> $GITHUB_ENV
 
-    - name: Test with pytest
-      run: pytest
-
-    - name: Create the package
-      run: python src/genescape/dist.py
-
-    - name: Upload resulting file as an artifact
-      uses: actions/upload-artifact@v3
+    - name: Upload artifact
+      uses: actions/upload-artifact@v4
       with:
-        name: GeneScape-latest.zip
+        name: GeneScape-${{ env.VERSION }}-Windows.zip
         path: dist/GeneScape-latest.zip
