+++ .github/workflows/draft-pdf.yml
@@ -1,3 +1,5 @@
+name: Generate JOSS paper
+
 on:
   push:
     paths:
