### 2024-04-03 594f46b
bugfix for cache, more information in nodes
