+++ src/genescape/utils.py
@@ -89,6 +89,9 @@ NODE_ATTRS = dict(fillcolor=BG_COLOR, shape=SHAPE, style="filled")
 # Namespace categories
 NS_BP, NS_MF, NS_CC, NS_ALL = "BP", "MF", "CC", "ALL"
 
+# Additional fields
+GO2GENE_COUNT, GO2GENE_PERC = "gene_count", "gene_percent"
+
 # Map the GO categories namespaces.
 NAMESPACE_MAP = {
     "biological_process": NS_BP,
