### 2024-04-03 84ce646
bugfix for cache, more information in nodes
