+++ .github/workflows/release-builder.yml
@@ -5,7 +5,8 @@ on:
   push:
     tags:
       - '*'
-
+  workflow_dispatch:
+  
 permissions:
   contents: read
 
