+++ .github/workflows/release-builder.yml
@@ -6,7 +6,7 @@ on:
     tags:
       - '*'
   workflow_dispatch:
-  
+
 permissions:
   contents: read
 
