+++ .github/workflows/draft-pdf.yml
@@ -5,7 +5,7 @@ on:
     paths:
       - 'docs/paper.*'
   workflow_dispatch:
-    
+
 jobs:
   paper:
     runs-on: ubuntu-latest
