+++ .github/workflows/release-builder.yml
@@ -27,7 +27,7 @@ jobs:
     - uses: actions/checkout@v4
 
     - name: Set up Python 3.10
-      uses: actions/setup-python@v4
+      uses: actions/setup-python@v5
       with:
         python-version: "3.10"
 
@@ -46,5 +46,5 @@ jobs:
     - name: Upload artifact
       uses: actions/upload-artifact@v4
       with:
-        name: GeneScape-${{ env.VERSION }}-${{ matrix.artifact-name }}.zip
+        name: GeneScape-${{ env.VERSION }}-${{ matrix.artifact-name }}
         path: dist/GeneScape-latest.zip
