### 2024-04-13 #6568 @j-andrews7
This looks interesting, and I'm happy to review within the next week.

I'll be out the week after, though that will allow Istvan the time to make any recommended changes before a second look :)
