+++ .github/workflows/release-builder.yml
@@ -37,7 +37,7 @@ jobs:
         pip install pytest pyinstaller
         pip install .
         pytest
-        python src/genescape/dist.py --build
+        python src/genescape/exe.py --build
 
     - name: Extract package version
       shell: bash
@@ -47,4 +47,4 @@ jobs:
       uses: actions/upload-artifact@v4
       with:
         name: GeneScape-${{ env.VERSION }}-${{ matrix.artifact-name }}
-        path: dist/GeneScape-latest.zip
+        path: dist/GeneScape-${{ matrix.artifact-name }}.zip
