### 2024-04-15 #6624 @editorialbot
[REVIEW]: GeneScape: A Python package for gene ontology analysis
**Submitting author:** @ialbert (<a href="http://orcid.org/0000-0001-8366-984X">Istvan Albert</a>)
**Repository:** https://github.com/ialbert/genescape-central
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.4
