+++ pyproject.toml
@@ -28,6 +28,7 @@ dependencies = [
   "toml" ,
   "networkx",
   "pydot",
+  "shiny",
 ]
 
 [project.urls]
@@ -64,7 +65,9 @@ packages = [
 [tool.hatch.version]
 path = "src/genescape/__about__.py"
 
+
 [tool.hatch.envs.default]
+path = "./venv"
 dependencies = [
   "coverage[toml]>=6.5",
   "pytest",
