+++ src/genescape/data/web/static/genescape.js
@@ -1,172 +1,0 @@
-
-// General error handler for HTMX events.
-function handleHTMXError(event) {
-    var err_msg = event.type + ': ' + (event.detail.errorMessage || 'server error')
-
-	// Show the error message in the console.
-	console.log(err_msg);
-
-	// Select the error layout elements.
-    var err_bar = document.getElementById('error-bar');
-	var err_div = document.getElementById('error-div');
-
-  	err_bar.style.display = 'block';
-
-    err_div.textContent = err_msg;
-    err_div.style.display = 'block';
-
-    // Hide the error message after a few seconds.
-    setTimeout(function() {
-        err_bar.style.display = 'none';
-        err_div.style.display = 'none';
-    }, 3000);
-
-}
-
-// Add event listeners for HTMX events.
-document.body.addEventListener('htmx:sendError', handleHTMXError);
-document.body.addEventListener('htmx:responseError', handleHTMXError);
-
-// Add event listener for the `data-toggle-width` attribute.
-document.addEventListener('click', function(e) {
-    if (e.target && e.target.dataset.toggleWidth === 'true') {
-        var side = document.getElementById('sideBar');
-        var main = document.getElementById('mainBar');
-
-        if (side.classList.contains('col-sm-3')) {
-            side.classList.remove('col-sm-3');
-            side.classList.add('col-sm-1');
-            main.classList.remove('col-sm-9');
-            main.classList.add('col-sm-11');
-        } else {
-            side.classList.remove('col-sm-1');
-            side.classList.add('col-sm-3');
-            main.classList.remove('col-sm-11');
-            main.classList.add('col-sm-9');
-        }
-    }
-});
-
-function submit(elemId, buttonId, eventName) {
-    const element = document.getElementById(elemId);
-    if (!element) return; // Exit if element does not exist
-
-    element.addEventListener(eventName, function(e) {
-        if (eventName === 'keydown' && e.key !== 'Enter') return;
-        e.preventDefault();
-        document.getElementById(buttonId).click();
-    });
-}
-
-document.addEventListener('DOMContentLoaded', (event) => {
-    submit('pattern', 'drawTree', 'keydown')
-    submit('count', 'drawTree', 'keydown')
-    submit('root', 'drawTree', 'change')
-});
-
-function resize_container(svg){
-
-	var svg = document.getElementById('svgtree')
-
-	var h = svg.getBoundingClientRect().height;
-	var w = svg.getBoundingClientRect().width;
-
-	var cn = document.getElementById('graphContainer');
-
-	cn.style.height = h + 'px';
-	cn.style.width  = w + 'px';
-
-	//console.log('cn:', cn.style.height, cn.style.width);
-}
-
-function render_graph(content) {
-	graph = document.getElementById('graph')
-	contr = document.getElementById('graphContainer')
-    Viz.instance().then(function(viz) {
-        var svg = viz.renderSVGElement(content);
-        svg.setAttribute("id", "svgtree");
-        //svg.setAttribute("width", "100%");
-        //svg.setAttribute("height", "1200px");
-        graph.innerHTML = '';
-        graph.appendChild(svg);
-
-        resize_container();
-
-    }).catch(
-        error => console.error("Error rendering Graphviz:", error)
-    );
-
-}
-
-var zoomLevel = 1; // Initial zoom level
-
-// Function to adjust SVG zoom based on action
-function adjustSVGZoom(zoomAction) {
-    var svg = document.querySelector('#graph svg');
-    if (!svg) return;
-
-	//console.log('Adjusting SVG zoom:', zoomAction);
-
-    // Adjust zoom level based on the action
-    if (zoomAction === 'zoom-in') {
[…]
