+++ src/genescape/web.py
@@ -3,7 +3,7 @@ from shiny import App, render, ui
 from pathlib import Path
 from datetime import datetime
 import time, asyncio
-from faicons import icon_svg
+from genescape import icons
 
 from genescape import __version__, annot, bottle, resources, tree, utils
 
@@ -34,12 +34,8 @@ digraph SimpleGraph {
 # Load the default resources.
 res = resources.init()
 
-# Load default icons
-icon_play = icon_svg("play")
-icon_down = icon_svg("download")
-icon_zoom_in = icon_svg("magnifying-glass-plus")
-icon_zoom_out = icon_svg("magnifying-glass-minus")
-zoom_reset = icon_svg("magnifying-glass")
+
+HOME = "https://github.com/ialbert/genescape-central/"
 
 # Shiny UI
 app_ui = ui.page_sidebar(
@@ -50,18 +46,18 @@ app_ui = ui.page_sidebar(
         ui.input_text("pattern", label="Word pattern (regex)", value="protein|cytoplasm"),
         ui.input_selectize(
             "selectize",
-            "Ontology root:",{
+            "Ontology root:", {
                 "AA": "All roots",
                 "BP": "Biological Process",
                 "MF": "Molecular Function",
                 "CC": "Cellular Component"},
         ),
-        ui.input_action_button("submit", "Draw Tree", class_="btn-success", icon=icon_play),
+        ui.input_action_button("submit", "Draw Tree", class_="btn-success", icon=icons.icon_play),
 
         ui.output_code("annot_elem"),
-        ui.download_link("download_annot", "Download annotations", icon=icon_down),
+        ui.download_link("download_csv", "Download annotations", icon=icons.icon_down),
         ui.output_code("dot_elem"),
-        ui.download_link("download_dot", "Download dot file", icon=icon_down),
+        ui.download_link("download_dot", "Download dot file", icon=icons.icon_down),
 
         width=400, bg="#f8f8f8",
     ),
@@ -81,22 +77,33 @@ app_ui = ui.page_sidebar(
     ),
 
     ui.tags.p(
-    ui.input_action_button("zoom_in", label="Zoom", icon=icon_zoom_in, class_="btn btn-light btn-sm",
-                   data_action="zoom-in"),
+        ui.input_action_button("zoom_in", label="Zoom", icon=icons.icon_zoom_in, class_="btn btn-light btn-sm",
+                               data_action="zoom-in"),
         ui.tags.span(" "),
-        ui.input_action_button("reset", label="Reset", icon=zoom_reset, class_="btn btn-light btn-sm", data_action="zoom-reset"),
+        ui.input_action_button("reset", label="Reset", icon=icons.zoom_reset, class_="btn btn-light btn-sm",
+                               data_action="zoom-reset"),
         ui.tags.span(" "),
-        ui.input_action_button("zoom_out", label="Zoom", icon=icon_zoom_out, class_="btn btn-light btn-sm", data_action="zoom-out"),
-ui.tags.span(" "),
-        ui.input_action_button("saveImage", "Save", class_="btn btn-light btn-sm", icon=icon_down),
-    align="center",
+        ui.input_action_button("zoom_out", label="Zoom", icon=icons.icon_zoom_out, class_="btn btn-light btn-sm",
+                               data_action="zoom-out"),
+        ui.tags.span(" "),
+        ui.input_action_button("saveImage", "Save", class_="btn btn-light btn-sm", icon=icons.icon_down),
+        align="center",
     ),
+
+    ui.tags.hr(),
 
     ui.tags.p(
         ui.div("""Press "Draw Tree" to generate the graph""", id="graph_elem", align="center"),
     ),
 
[…]
