+++ pyproject.toml
@@ -87,7 +87,7 @@ cov = [
 ]
 
 [[tool.hatch.envs.all.matrix]]
-python = ["3.10", "3.11", "3.12"]
+python = ["3.10", "3.11"]
 
 [tool.hatch.envs.lint]
 detached = true
