+++ src/genescape/web.py
@@ -5,24 +5,20 @@ from datetime import datetime
 import time, asyncio
 from genescape import icons
 
-from genescape import __version__, annot, bottle, resources, tree, utils
+from genescape import __version__, annot, build, resources, tree, utils
 
+# Default gene list.
 GENE_LIST = '''
 ABTB3
 BCAS4
 C3P1
 GRTP1
-ABTB3
-BCAS4
-C3P1
-GRTP1
-Cyp1a1
-Sphk2
-Sptlc2
-GO:0005488
-GO:0005515
 '''
 
+# Default pattern.
+PATTERN = ""
+
+# Default dot file.
 DOT = '''
 digraph SimpleGraph {
     A -> B;
@@ -31,11 +27,20 @@ digraph SimpleGraph {
 }
 '''
 
+# Default sidebar width.
+SIDEBAR_WIDTH = 300
+
+# Default sidebar background color.
+SIDEBAR_BG = "#f8f8f8"
+
+# Home page link
+HOME = "https://github.com/ialbert/genescape-central/"
+
 # Load the default resources.
 res = resources.init()
 
-
-HOME = "https://github.com/ialbert/genescape-central/"
+# This is the global index name
+INDEX = res.INDEX
 
 # Shiny UI
 app_ui = ui.page_sidebar(
@@ -43,14 +48,15 @@ app_ui = ui.page_sidebar(
     ui.sidebar(
         ui.input_text_area("terms", label="Gene List", value=GENE_LIST),
         ui.input_text("mincount", label="Minimum count (integer)", value="1"),
-        ui.input_text("pattern", label="Word pattern (regex)", value="protein|cytoplasm"),
-        ui.input_selectize(
-            "selectize",
+        ui.input_text("pattern", label="Word pattern (regex)", value=PATTERN),
+        ui.input_select(
+            "root",
             "Ontology root:", {
-                "AA": "All roots",
-                "BP": "Biological Process",
-                "MF": "Molecular Function",
-                "CC": "Cellular Component"},
+                utils.NS_ALL: "All roots",
+                utils.NS_BP: "Biological Process",
+                utils.NS_MF: "Molecular Function",
+                utils.NS_CC: "Cellular Component"
+            },
         ),
         ui.input_action_button("submit", "Draw Tree", class_="btn-success", icon=icons.icon_play),
 
@@ -59,7 +65,7 @@ app_ui = ui.page_sidebar(
         ui.output_code("dot_elem"),
         ui.download_link("download_dot", "Download dot file", icon=icons.icon_down),
 
-        width=400, bg="#f8f8f8",
+        width=SIDEBAR_WIDTH, bg=SIDEBAR_BG,
     ),
 
     ui.head_content(
@@ -106,7 +112,7 @@ app_ui = ui.page_sidebar(
     ),
     ui.output_text("run_elem"),
 
-    title="GeneScape", id="main",
+    title=f"GeneScape" , id="main",
 
 )
 
@@ -128,18 +134,20 @@ def server(input, output, session):
     ann_value = reactive.Value("# The annotations will appear here.")
     dot_value = reactive.Value("# The dot file will appear here.")
 
+    res = resources.init()
+
+    # External index or default
+    index = INDEX or res.INDEX
+
     async def create_tree(text):
         mincount = int(input.mincount())
         pattern = input.pattern()
 
         inp = text2list(text)
 
-        cnf = resources.get_config()
+        root = input.root()
 
-        res = resources.init(cnf)
-        index = res.INDEX
-
-        graph, ann = tree.parse_input(inp=inp, index=index, mincount=mincount, pattern=pattern)
+        graph, ann = tree.parse_input(inp=inp, index=index, mincount=mincount, pattern=pattern, root=root)
 
         dot = tree.write_tree(graph, ann, out=None)
 
[…]
