+++ src/genescape/bottle.py
@@ -1,4417 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-"""
-Bottle is a fast and simple micro-framework for small web applications. It
-offers request dispatching (Routes) with URL parameter support, templates,
-a built-in HTTP Server and adapters for many third party WSGI/HTTP-server and
-template engines - all in a single file and with no dependencies other than the
-Python Standard Library.
-
-Homepage and documentation: http://bottlepy.org/
-
-Copyright (c) 2009-2018, Marcel Hellkamp.
-License: MIT (see LICENSE for details)
-"""
-
-from __future__ import print_function
-import sys
-
-__author__ = 'Marcel Hellkamp'
-__version__ = '0.13-dev'
-__license__ = 'MIT'
-
-###############################################################################
-# Command-line interface ######################################################
-###############################################################################
-# INFO: Some server adapters need to monkey-patch std-lib modules before they
-# are imported. This is why some of the command-line handling is done here, but
-# the actual call to _main() is at the end of the file.
-
-
-def _cli_parse(args):  # pragma: no coverage
-    from argparse import ArgumentParser
-
-    parser = ArgumentParser(prog=args[0], usage="%(prog)s [options] package.module:app")
-    opt = parser.add_argument
-    opt("--version", action="store_true", help="show version number.")
-    opt("-b", "--bind", metavar="ADDRESS", help="bind socket to ADDRESS.")
-    opt("-s", "--server", default='wsgiref', help="use SERVER as backend.")
-    opt("-p", "--plugin", action="append", help="install additional plugin/s.")
-    opt("-c", "--conf", action="append", metavar="FILE",
-        help="load config values from FILE.")
-    opt("-C", "--param", action="append", metavar="NAME=VALUE",
-        help="override config values.")
-    opt("--debug", action="store_true", help="start server in debug mode.")
-    opt("--reload", action="store_true", help="auto-reload on file changes.")
-    opt('app', help='WSGI app entry point.', nargs='?')
-
-    cli_args = parser.parse_args(args[1:])
-
-    return cli_args, parser
-
-
-def _cli_patch(cli_args):  # pragma: no coverage
-    parsed_args, _ = _cli_parse(cli_args)
-    opts = parsed_args
-    if opts.server:
-        if opts.server.startswith('gevent'):
-            import gevent.monkey
-            gevent.monkey.patch_all()
-        elif opts.server.startswith('eventlet'):
-            import eventlet
-            eventlet.monkey_patch()
-
-
-if __name__ == '__main__':
-    _cli_patch(sys.argv)
-
-###############################################################################
-# Imports and Python 2/3 unification ##########################################
-###############################################################################
-
-import base64, calendar, cgi, email.utils, functools, hmac, itertools,\
-       mimetypes, os, re, tempfile, threading, time, warnings, weakref, hashlib
-
-from types import FunctionType
-from datetime import date as datedate, datetime, timedelta
-from tempfile import NamedTemporaryFile
-from traceback import format_exc, print_exc
-from unicodedata import normalize
-
-try:
-    from ujson import dumps as json_dumps, loads as json_lds
-except ImportError:
-    from json import dumps as json_dumps, loads as json_lds
-
-py = sys.version_info
-py3k = py.major > 2
-
-# Lots of stdlib and builtin differences.
-if py3k:
-    import http.client as httplib
-    import _thread as thread
-    from urllib.parse import urljoin, SplitResult as UrlSplitResult
-    from urllib.parse import urlencode, quote as urlquote, unquote as urlunquote
-    urlunquote = functools.partial(urlunquote, encoding='latin1')
-    from http.cookies import SimpleCookie, Morsel, CookieError
[…]
