+++ docs/paper.bib
@@ -92,3 +92,13 @@
     eprint = {https://academic.oup.com/bioinformatics/article-pdf/25/22/3045/48997998/bioinformatics\_25\_22\_3045.pdf},
 }
 
+@InProceedings{networkx,
+  author =       {Aric A. Hagberg and Daniel A. Schult and Pieter J. Swart},
+  title =        {Exploring Network Structure, Dynamics, and Function using NetworkX},
+  booktitle =   {Proceedings of the 7th Python in Science Conference},
+  pages =     {11 - 15},
+  address = {Pasadena, CA USA},
+  year =      {2008},
+  editor =    {Ga\"el Varoquaux and Travis Vaught and Jarrod Millman},
+}
+
