+++ src/genescape/shiny/app.py
@@ -43,78 +43,82 @@ res = resources.init()
 # This is the global index name
 INDEX = res.INDEX
 
-app_ui = ui.page_sidebar(
+def generate_ui():
 
-    ui.sidebar(
-        ui.input_text_area("terms", label="Gene List", value=GENE_LIST),
-        ui.input_text("mincount", label="Minimum count (integer)", value="1"),
-        ui.input_text("pattern", label="Word pattern (regex)", value=PATTERN),
-        ui.input_select(
-            "root",
-            "Ontology root:", {
-                utils.NS_ALL: "All roots",
-                utils.NS_BP: "Biological Process",
-                utils.NS_MF: "Molecular Function",
-                utils.NS_CC: "Cellular Component"
-            },
+    app_ui = ui.page_sidebar(
+
+        ui.sidebar(
+            ui.input_text_area("terms", label="Gene List", value=GENE_LIST),
+            ui.input_text("mincount", label="Minimum count (integer)", value="1"),
+            ui.input_text("pattern", label="Word pattern (regex)", value=PATTERN),
+            ui.input_select(
+                "root",
+                "Ontology root:", {
+                    utils.NS_ALL: "All roots",
+                    utils.NS_BP: "Biological Process",
+                    utils.NS_MF: "Molecular Function",
+                    utils.NS_CC: "Cellular Component"
+                },
+            ),
+            ui.input_action_button("submit", "Draw Tree", class_="btn-success", icon=icons.icon_play),
+
+            ui.output_code("annot_elem"),
+            ui.download_link("download_csv", "Download annotations", icon=icons.icon_down),
+            ui.output_code("dot_elem"),
+            ui.download_link("download_dot", "Download dot file", icon=icons.icon_down),
+
+            width=SIDEBAR_WIDTH, bg=SIDEBAR_BG,
         ),
-        ui.input_action_button("submit", "Draw Tree", class_="btn-success", icon=icons.icon_play),
 
-        ui.output_code("annot_elem"),
-        ui.download_link("download_csv", "Download annotations", icon=icons.icon_down),
-        ui.output_code("dot_elem"),
-        ui.download_link("download_dot", "Download dot file", icon=icons.icon_down),
-
-        width=SIDEBAR_WIDTH, bg=SIDEBAR_BG,
-    ),
-
-    ui.head_content(
-        ui.tags.script(
-            """
-            $(function() {
-                Shiny.addCustomMessageHandler("trigger", function(message) {
-                    render_graph_delay();
+        ui.head_content(
+            ui.tags.script(
+                """
+                $(function() {
+                    Shiny.addCustomMessageHandler("trigger", function(message) {
+                        render_graph_delay();
+                    });
                 });
-            });
-            """),
-        ui.include_css(res.GENESCAPE_CSS),
-        ui.include_js(res.VIZ_JS, defer=""),
-        ui.include_js(res.GENESCAPE_JS, method="inline"),
-    ),
-
-    ui.tags.p(
-        ui.input_action_button("zoom_in", label="Zoom", icon=icons.icon_zoom_in, class_="btn btn-light btn-sm",
-                               data_action="zoom-in"),
-        ui.tags.span(" "),
-        ui.input_action_button("reset", label="Reset", icon=icons.zoom_reset, class_="btn btn-light btn-sm",
-                               data_action="zoom-reset"),
-        ui.tags.span(" "),
-        ui.input_action_button("zoom_out", label="Zoom", icon=icons.icon_zoom_out, class_="btn btn-light btn-sm",
[…]
