+++ src/genescape/main.py
@@ -3,6 +3,7 @@ from pathlib import Path
 from genescape import __version__
 from genescape import resources, utils
 import shiny
+import os
 
 
 HELP  = f"Gene function visualization (v{__version__})."
@@ -178,16 +179,20 @@ port= 8000,
 @cli.command()
 @click.option("--host", "host", default="127.0.0.1", help="hostname to bind to")
 @click.option("--port", "port", default=8000, type=int, help="port number")
+@click.option("-i", "--index", "index", help="Genescape index file")
 @click.option("--reload", "reload", is_flag=True, help="reload the server on changes")
 @click.option("--reset", "reset", is_flag=True, help="resets resources")
-@click.option("-i", "--index", "index", help="Genescape index file")
-
+@click.option("-c", "--conf", "conf", help="pass a TOML config file")
 @click.option("-v", "verbose", is_flag=True, help="Verbose output.")
 @click.help_option("-h", "--help")
-def web(index=None, host=None, port=None, reload=False, reset=False, verbose=False):
+def web(index=None, host=None, port=None, reload=False, reset=False, conf=None, verbose=False):
     """
     Run the web interface.
     """
+
+    # TODO: Implement the configuration file.
+    if conf:
+        utils.stop(f"Config not yet implemented")
 
     # Reset the resource directory
     if reset:
@@ -196,12 +201,13 @@ def web(index=None, host=None, port=None, reload=False, reset=False, verbose=Fal
     # Set the verbosity level.
     utils.verbosity(verbose)
 
-    cnf = resources.get_config()
-
-    res = resources.init(cnf)
-
-    # Set the generack index file.
-    #web.INDEX = Path(index) if index else res.INDEX
+    # Insert the index into the environment.
+    if index:
+        if not os.path.isfile(index):
+            utils.stop(f"# Index not found: {index}")
+        key = "idx"
+        label = str(Path(index).name).split(".")[0].title()
+        os.environ['GENESCAPE_INDEX'] = f'{key}:{label}:{index}'
 
     # Run the server.
     shiny.run_app("genescape.shiny.app:app", host=host, port=port, reload=reload, factory=True)
