+++ src/genescape/shiny/app.py
@@ -1,6 +1,6 @@
 from shiny import reactive
 from shiny import App, render, ui
-import  asyncio, os
+import asyncio, os
 from genescape import icons
 from genescape import __version__, annot, resources, tree, utils
 
@@ -42,91 +42,88 @@ SIDEBAR_BG = res.config.get("SIDEBAR_WIDTH", "#f9f9f9")
 # Home page link
 HOME = "https://github.com/ialbert/genescape-central/"
 
-def generate_ui():
-    app_ui = ui.page_sidebar(
+app_ui = ui.page_sidebar(
 
-        ui.sidebar(
-            ui.input_text_area("terms", label="Gene List", value=GENE_LIST),
+    ui.sidebar(
+        ui.input_text_area("terms", label="Gene List", value=GENE_LIST),
 
-            ui.div({"class": "left-aligned"},
+        ui.div({"class": "left-aligned"},
 
-                   ui.input_select(
-                       "database",
-                       "Organism", DATABASE_CHOICES,
-                   ),
+               ui.input_select(
+                   "database",
+                   "Organism", DATABASE_CHOICES,
+               ),
 
-                   ui.input_text("mincount", label="Mincount", value=MINCOUNT),
-                   ui.input_text("pattern", label="Pattern", value=PATTERN),
+               ui.input_text("mincount", label="Mincount", value=MINCOUNT),
+               ui.input_text("pattern", label="Pattern", value=PATTERN),
 
-                   ui.input_select(
-                       "root",
-                       "Root", {
-                           utils.NS_ALL: "All roots",
-                           utils.NS_BP: "Biological Process",
-                           utils.NS_MF: "Molecular Function",
-                           utils.NS_CC: "Cellular Component"
-                       },
-                   ),
-                   ),
-            ui.input_action_button("submit", "Draw Tree", class_="btn-success", icon=icons.icon_play),
+               ui.input_select(
+                   "root",
+                   "Root", {
+                       utils.NS_ALL: "All roots",
+                       utils.NS_BP: "Biological Process",
+                       utils.NS_MF: "Molecular Function",
+                       utils.NS_CC: "Cellular Component"
+                   },
+               ),
+               ),
+        ui.input_action_button("submit", "Draw Tree", class_="btn-success", icon=icons.icon_play),
 
-            ui.output_code("annot_elem"),
-            ui.download_link("download_csv", "Download annotations", icon=icons.icon_down),
-            ui.output_code("dot_elem"),
-            ui.download_link("download_dot", "Download dot file", icon=icons.icon_down),
+        ui.output_code("annot_elem"),
+        ui.download_link("download_csv", "Download annotations", icon=icons.icon_down),
+        ui.output_code("dot_elem"),
+        ui.download_link("download_dot", "Download dot file", icon=icons.icon_down),
 
-            width=SIDEBAR_WIDTH, bg=SIDEBAR_BG,
-        ),
+        width=SIDEBAR_WIDTH, bg=SIDEBAR_BG,
+    ),
 
-        ui.head_content(
-            ui.tags.script(
-                """
-                $(function() {
-                    Shiny.addCustomMessageHandler("trigger", function(message) {
-                        render_graph();
-                    });
+    ui.head_content(
+        ui.tags.script(
+            """
+            $(function() {
+                Shiny.addCustomMessageHandler("trigger", function(message) {
+                    render_graph();
                 });
-                """),
-            ui.include_css(res.GENESCAPE_CSS),
[…]
