+++ Makefile
@@ -31,7 +31,7 @@ usage:
 
 # Performs the testing.
 web:
-	shiny run --factory --reload genescape.shiny.app:app
+	shiny run --reload genescape.shiny.app:app
 
 # Tag and push to repository.
 tag: test
@@ -104,6 +104,7 @@ env:
 	micromamba create -n shiny python=3.11 rsconnect-python graphviz make -y
 
 shiny:
+	#pip install rsconnect-python
 	rsconnect deploy shiny src/genescape/shiny --name biostar --title GeneScape
 
 .PHONY: test lint fix push clean build publish obo docimg web
