+++ docs/paper.bib
@@ -102,3 +102,10 @@
   editor =    {Ga\"el Varoquaux and Travis Vaught and Jarrod Millman},
 }
 
+@Manual{shiny,
+  title = {shiny: Web Application Framework for R},
+  author = {Winston Chang and Joe Cheng and JJ Allaire and Carson Sievert and Barret Schloerke and Yihui Xie and Jeff Allen and Jonathan McPherson and Alan Dipert and Barbara Borges},
+  year = {2024},
+  note = {R package version 1.8.1.9000, https://github.com/rstudio/shiny},
+  url = {https://shiny.posit.co/},
+}
