### 2024-04-22 #6624 @ialbert
@j-andrews7 @sridhar0605

Based on reviewer suggestions, the software has been rewritten to be deployed as Shiny app and is now also available at

* https://biostar.shinyapps.io/genescape/

A few minor alterations were made to the paper, citing Networkx and Shiny, additional screenshot of the Shiny interface was also added. All other suggestions were addressed.
