### 2024-04-25 4eb2cb4
work on alternative representation
