+++ src/genescape/graph.py
@@ -1,0 +1,370 @@
+"""
+Graph representation of an ontology
+"""
+
+import gzip, sys, re, json, csv, textwrap, pickle
+from genescape import resources, utils
+from itertools import islice, takewhile, dropwhile, tee, chain
+import networkx as nx
+
+# Namespace categories
+NS_BP, NS_MF, NS_CC, NS_ALL = "BP", "MF", "CC", "ALL"
+
+# Additional fields
+OUT_DEGREE = "out_degree"
+IN_DEGREE = "in_degree"
+
+# Map the GO categories namespaces.
+NAMESPACE_MAP = {
+    "biological_process": NS_BP,
+    "molecular_function": NS_MF,
+    "cellular_component": NS_CC,
+}
+
+# Set the default attributes for the nodes
+# NODE_ATTRS = dict(fillcolor=BG_COLOR, shape=SHAPE, style="filled")
+
+NODE_ATTRS = {OUT_DEGREE: 0}
+
+HEADER_KEY = "header"
+TERMS_KEY = "terms"
+TYPEDEFS_KEY = "typedefs"
+INPUT_LIST = "input_list"
+GENE_LIST = "gene_list"
+GENE_LEN = "gene_len"
+
+SINGLE_VALUE = {"id", "name", "namespace", "comment"}
+MULTI_VALUE = {"alt_id", "subset", "replaced_by", "consider"}
+
+SYM2GO, GO2SYM, GO2NAME = "sym2go", "go2sym", "go2name"
+
+
+@utils.timer
+def build_graph(data):
+    # The complete ontology graph
+    graph = nx.DiGraph()
+    terms = data[TERMS_KEY]
+
+    nodes = filter(lambda x: not x.get("is_obsolete"), terms.values())
+    # nodes = islice(nodes, 100)
+    nodes = list(nodes)
+
+    # Add all the nodes
+    for node in nodes:
+        oid = node["id"]
+        name = node["name"]
+        namespace = utils.NAMESPACE_MAP.get(node["namespace"], "?")
+        graph.add_node(oid, id=oid, name=name, namespace=namespace, **utils.NODE_ATTRS)
+
+    for node in nodes:
+        child = node["id"]
+        for elem in node.get("is_a", []):
+            parent = elem["target"]
+            if graph.has_node(parent):
+                graph.add_edge(parent, child)
+            else:
+                utils.warn(f"# Missing parent: {parent} for {node}")
+
+    # Add the degree to the nodes
+    for node in graph.nodes():
+        out_degree = graph.out_degree(node)
+        in_degree = graph.in_degree(node)
+        graph.nodes[node][OUT_DEGREE] = out_degree
+        graph.nodes[node][IN_DEGREE] = in_degree
+
+    # save with cpickle
+    with open("../../graph.pkl", "wb") as fp:
+        pickle.dump(graph, fp)
+
+
+@utils.timer
+def parse_gaf(fname):
+    stream = gzip.open(fname, mode="rt", encoding="UTF-8")
+    stream = filter(lambda x: not x.startswith("!"), stream)
+    stream = map(lambda x: x.strip(), stream)
+    stream = csv.reader(stream, delimiter="\t")
+    # stream = islice(stream, 1000)
+    sym2go, go2sym, go2name = {}, {}, {}
+    for row in stream:
+        name, symb, goid, synon = row[1], row[2], row[4], row[10]
+
+        sym2go.setdefault(symb, []).append(goid)
+        sym2go.setdefault(name, []).append(goid)
+
+        go2sym.setdefault(goid, []).append(symb)
+        go2name.setdefault(goid, []).append(name)
+
+    # Remove duplicates in case these exists.
[…]
