### 2024-04-25 7b9b928
work on alternative representation
