### 2024-04-26 e9ad041
work on new representation
