+++ src/genescape/graph.py
@@ -26,59 +26,68 @@ NAMESPACE_MAP = {
 # Set the default attributes for the nodes
 # NODE_ATTRS = dict(fillcolor=BG_COLOR, shape=SHAPE, style="filled")
 
-NODE_ATTRS = {OUT_DEGREE: 0}
 
-HEADER_KEY = "header"
-TERMS_KEY = "terms"
-TYPEDEFS_KEY = "typedefs"
+# Data keys in the index.
+HEADER_KEY = "info"
+
+# Ontology key.
+OBO_KEY = "obo"
+
+# Mapping keys.
+SYM2GO, GO2SYM, NAME2SYM = "sym2go", "go2sym", "name2sym"
+
+
 INPUT_LIST = "input_list"
-GENE_LIST = "gene_list"
-GENE_LEN = "gene_len"
-TOTAL_COUNT = "total_count"
+SOURCE_SYM = "sources"
+SOURCE_LEN = "src_len"
 
-SINGLE_VALUE = {"id", "name", "namespace", "comment"}
-MULTI_VALUE = {"alt_id", "subset", "replaced_by", "consider"}
+# The number of descendants for the GO term.
+DESC_COUNT = "desc_count"
 
-SYM2GO, GO2SYM, GO2NAME = "sym2go", "go2sym", "go2name"
+# The number of annotation for the GO term.
+ANNO_COUNT = "anno_count"
 
+# The cumulative sum of annotations for GO term.
+ANNO_TOTAL = "anno_total"
+
+# Node attributes
+NODE_ATTRS = {OUT_DEGREE: 0, DESC_COUNT: 0, ANNO_COUNT: 0, ANNO_TOTAL: 0}
 
 @utils.timer
 def build_graph(data):
+    """
+    Build an ontology graph from a JSON data structure.
+    """
     # The complete ontology graph
     graph = nx.DiGraph()
-    terms = data[TERMS_KEY]
+    terms = data[OBO_KEY]
 
-    nodes = filter(lambda x: not x.get("is_obsolete"), terms.values())
+    terms = filter(lambda x: not x.get("is_obsolete"), terms.values())
     # nodes = islice(nodes, 100)
-    nodes = list(nodes)
+    terms = list(terms)
 
     # Add individual nodes to the
-    for node in nodes:
-        oid = node["id"]
-        name = node["name"]
-        namespace = utils.NAMESPACE_MAP.get(node[NAMESPACE], "?")
-        graph.add_node(oid, id=oid, name=name, namespace=namespace, **utils.NODE_ATTRS)
+    for row in terms:
+        oid = row["id"]
+        name = row["name"]
+        namespace = utils.NAMESPACE_MAP.get(row[NAMESPACE], "?")
+        anno_count = row.get(ANNO_COUNT, -1)
+        anno_total = row.get(ANNO_TOTAL, -1)
+        desc_count = row.get(DESC_COUNT, -1)
+        out_degree = row.get(OUT_DEGREE, -1)
+        in_degree = row.get(IN_DEGREE, -1)
+        kwargs = {OUT_DEGREE: out_degree, IN_DEGREE: in_degree, DESC_COUNT: desc_count, ANNO_COUNT: anno_count, ANNO_TOTAL: anno_total}
+        graph.add_node(oid, id=oid, name=name, namespace=namespace, **kwargs)
 
-    for node in nodes:
-        child = node["id"]
-        for elem in node.get("is_a", []):
-            parent = elem["target"]
+    for row in terms:
+        child = row["id"]
+        for parent in row.get("is_a", []):
             if graph.has_node(parent):
                 graph.add_edge(parent, child)
             else:
-                utils.warn(f"# Missing parent: {parent} for {node}")
[…]
