### 2024-04-26 8e3d1f7
work on new representation
