+++ src/genescape/graph.py
@@ -36,10 +36,17 @@ OBO_KEY = "obo"
 # Mapping keys.
 SYM2GO, GO2SYM, NAME2SYM = "sym2go", "go2sym", "name2sym"
 
+# The ID is in the input
+SRC_FLAG = "src_flag"
 
-INPUT_LIST = "input_list"
-SOURCE_SYM = "sources"
-SOURCE_LEN = "src_len"
+# The symbols in the input
+SRC_SYMB = "src_sym"
+
+# The number of symbols in the input
+SRC_SIZE = "src_len"
+
+# The total number of symbols in the input
+SRC_TOT = "src_tot"
 
 # The number of descendants for the GO term.
 DESC_COUNT = "desc_count"
@@ -198,6 +205,10 @@ def parse_obo(stream):
 @utils.timer
 def make_subgraph(idx, graph, tgt=None, root=NS_MF):
 
+    #tgt="GO:0035639 GO:0035639 GO:0097159  ".split()
+
+    #tgt = "GO:0035639 GO:0035639 GO:0043168 GO:0043168 GO:0043168 GO:0097159 GO:0005488  GO:1901265  GO:0035639  GO:0035639".split()
+
     tgt = tgt or "ABTB3 BCAS4 C3P1 GRTP1 SPOP ABCA2 PSMD3 GO:0046983 GO:1901265".split()
 
     # Unique targets
@@ -214,9 +225,15 @@ def make_subgraph(idx, graph, tgt=None, root=NS_MF):
     valid2sym = list(filter(lambda x: x in sym2go or x in obo, tgt))
 
     # Map the symbols to GO terms
-    goids = map(lambda x: sym2go.get(x, [x]), valid2sym)
-    goids = chain.from_iterable(goids)
-    goids = list(set(goids))
+    goids, go2tgt = [], dict()
+    for sym in valid2sym:
+        values = sym2go.get(sym, []) or [sym]
+        for goid in values:
+            go2tgt.setdefault(goid, []).append(sym)
+            goids.append(goid)
+
+    # Map the symbols to GO terms
+
 
     # Terms not found in the OBO
     miss2go = set(filter(lambda x: x not in graph.nodes, goids))
@@ -242,14 +259,28 @@ def make_subgraph(idx, graph, tgt=None, root=NS_MF):
     # Initialize the subtree
     for node_id in tree.nodes():
         node = tree.nodes[node_id]
-        is_input = node_id in valid2tgt
-        if is_input:
-            sources = list(set(go2sym.get(node_id, [])) & tgt2set)
-        else:
-            sources = []
-        node[INPUT_LIST] = is_input
-        node[SOURCE_SYM] = sources
-        node[SOURCE_LEN] = len(sources)
+        srcs = go2tgt.get(node_id, [])
+        node[SRC_FLAG] = node_id in valid2tgt
+        node[SRC_SYMB] = list(srcs)
+        node[SRC_SIZE] = len(srcs)
+
+    # Travers and update the cumulative totals
+    def traverse(tree, node_id, store):
+        store[node_id] = [ node_id ]
+        for child in tree.successors(node_id):
+            if child not in store:
+                traverse(tree, child, store)
+                store[node_id].extend(store[child])
+
+        src_sym = []
+        for sub_id in store[node_id]:
+            src_sym += tree.nodes[sub_id][SRC_SYMB]
+        tree.nodes[node_id][SRC_TOT] = len(src_sym)
+
+    roots = filter(lambda x: tree.in_degree(x) == 0, tree.nodes())
+    for root in roots:
+        store = dict()
+        traverse(tree, root, store=store)
 
     # Sort nodes
     s_nodes = sorted(tree.nodes(data=True))
[…]
