+++ src/genescape/graph.py
@@ -1,6 +1,8 @@
 """
 Graph representation of an ontology
 """
+import click
+from genescape import cli
 
 import gzip, sys, re, json, csv, textwrap, pickle
 from genescape import resources, utils
@@ -36,29 +38,50 @@ OBO_KEY = "obo"
 # Mapping keys.
 SYM2GO, GO2SYM, NAME2SYM = "sym2go", "go2sym", "name2sym"
 
-# The ID is in the input
-SRC_FLAG = "src_flag"
+# True for a GO term that is present in the input data.
+INP_FLAG = "inp_flag"
 
-# The symbols in the input
-SRC_SYMB = "src_sym"
+# A list that stores the symbols that support a GO term.
+INP_SYM = "inp_sym"
 
-# The number of symbols in the input
-SRC_SIZE = "src_len"
+# The length of the INP_SYMB list.
+INP_LEN = "inp_len"
 
-# The total number of symbols in the input
-SRC_TOT = "src_tot"
+# A list that stores all the of a GO subtree.
+INP_SYM_TOT = "inp_sym_tot"
+
+# The lenght of the INP_SYMB_TOT list.
+INP_LEN_TOT = "inp_len_tot"
+
+# The list of descendant nodes for the GO term.
+DESC_LIST = "desc_list"
 
 # The number of descendants for the GO term.
 DESC_COUNT = "desc_count"
 
-# The number of annotation for the GO term.
+# The number of annotations for the GO term.
 ANNO_COUNT = "anno_count"
 
 # The cumulative sum of annotations for GO term.
 ANNO_TOTAL = "anno_total"
 
+# Status keys
+SYM_VALID = "sym_valid"
+SYM_MISS = "sym_miss"
+GO_VALID = "go_valid"
+GO_MISS = "go_miss"
+
 # Node attributes
 NODE_ATTRS = {OUT_DEGREE: 0, DESC_COUNT: 0, ANNO_COUNT: 0, ANNO_TOTAL: 0}
+
+WIDTH = 16.9
+HEIGHT = 6.0
+
+
+def dpi(size):
+    value = size / WIDTH
+    return value
+
 
 @utils.timer
 def build_graph(data):
@@ -83,7 +106,8 @@ def build_graph(data):
         desc_count = row.get(DESC_COUNT, -1)
         out_degree = row.get(OUT_DEGREE, -1)
         in_degree = row.get(IN_DEGREE, -1)
-        kwargs = {OUT_DEGREE: out_degree, IN_DEGREE: in_degree, DESC_COUNT: desc_count, ANNO_COUNT: anno_count, ANNO_TOTAL: anno_total}
+        kwargs = {OUT_DEGREE: out_degree, IN_DEGREE: in_degree, DESC_COUNT: desc_count, ANNO_COUNT: anno_count,
+                  ANNO_TOTAL: anno_total}
         graph.add_node(oid, id=oid, name=name, namespace=namespace, **kwargs)
 
     for row in terms:
@@ -95,6 +119,7 @@ def build_graph(data):
                 utils.warn(f"# Missing parent: {parent} for {row}")
 
     return graph
+
 
 @utils.timer
 def parse_gaf(fname):
@@ -128,6 +153,7 @@ def parse_gaf(fname):
 
     return gaf
 
+
 def parse_refs(text):
     text = text.strip('[').strip(']').strip()
     refs = [ref for ref in text.split(",") if ref]
@@ -157,7 +183,7 @@ def parse_obo(stream):
 
     data = {HEADER_KEY: head, OBO_KEY: terms}
 
-    INCLUDE = [ "format-version", "data-version", "ontology" ]
[…]
