+++ src/genescape/graph.py
@@ -1,9 +1,8 @@
 """
 Graph representation of an ontology
 """
-import click
-from genescape import cli
-
+import io
+from pathlib import Path
 import gzip, sys, re, json, csv, textwrap, pickle
 from genescape import resources, utils
 from itertools import islice, takewhile, dropwhile, tee, chain
@@ -83,7 +82,6 @@ def dpi(size):
     return value
 
 
-@utils.timer
 def build_graph(data):
     """
     Build an ontology graph from a JSON data structure.
@@ -226,16 +224,10 @@ def parse_obo(stream):
 
     return data
 
+def subgraph(idx, graph, tgt, root=NS_MF, mincount=1, pattern=''):
 
-@utils.timer
-def make_subgraph(idx, graph, tgt=[], root=NS_MF, mincount=1, pattern=''):
-    # tgt="GO:0035639 GO:0035639 GO:0097159  ".split()
-
-    # tgt = "GO:0035639 GO:0035639 GO:0043168 GO:0043168 GO:0043168 GO:0097159 GO:0005488  GO:1901265  GO:0035639  GO:0035639".split()
-
-    tgt = "GO:1901265 GO:0097159".split()
-
-    tgt = tgt or "ABTB3 BCAS4 C3P1 GRTP1 SPOP ABCA2 PSMD3 GO:0046983 GO:1901265".split()
+    # Better defaults
+    tgt = tgt or []
 
     # Inputs must be unique.
     tgt = set(tgt)
@@ -244,7 +236,7 @@ def make_subgraph(idx, graph, tgt=[], root=NS_MF, mincount=1, pattern=''):
     sym2go = idx[SYM2GO]
 
     # Symbols not found in the association file.
-    sym_miss = set(filter(lambda x: x not in sym2go, tgt))
+    sym_miss = set(filter(lambda x: x not in sym2go and x not in obo, tgt))
 
     # Print notification on missing symbols.
     if sym_miss:
@@ -374,10 +366,18 @@ def human_readable(value, digits=0):
     newval = f"{newval:d}K" if newval > 1 else value
     return newval
 
+@utils.memoize
+def load_json(path):
 
-@utils.timer
-def load_json(fname):
-    stream = gzip.open(fname, "rb") if fname.endswith(".gz") else open(fname, "rt")
+    if not isinstance(path, Path):
+        path = Path(path)
+
+
+    # Print error if path does not exist
+    if not path.exists():
+        utils.stop(f"file not found: {path}")
+
+    stream = gzip.open(path, "rb") if path.name.endswith(".gz") else open(path, "rt")
     with stream as fp:
         text = fp.read().decode('utf-8')
         data = json.loads(text)
@@ -386,15 +386,17 @@ def load_json(fname):
 
 
 @utils.timer
-def save_json(data, fname):
-    stream = gzip.open(fname, "wb") if fname.endswith(".gz") else open(fname, "wb")
+def save_json(data, path):
+    if not isinstance(path, Path):
+        path = Path(path)
+    stream = gzip.open(path, "wb") if path.name.endswith(".gz") else open(path, "wb")
     with stream as fp:
         text = json.dumps(data, indent=4)
         fp.write(text.encode('utf-8'))
 
 
 @utils.timer
-def compute_graph_measures(idx):
+def fill_graph(idx):
     """
     Fills the index with additional information
     """
@@ -431,8 +433,6 @@ def compute_graph_measures(idx):
 
     return obo
 
-
-@utils.memoize
 def load_index(fname):
     idx = load_json(fname)
[…]
