+++ src/genescape/main.py
@@ -1,218 +1,113 @@
-import click
-from pathlib import Path
 from genescape import __version__
 from genescape import resources, utils
-import shiny
-import os
+from genescape import graph
+import click, csv
+import sys, json
+from pathlib import Path
+
+# Valid choices for root
+ROOT_CHOICES = [utils.NS_BP, utils.NS_MF, utils.NS_CC, utils.NS_ALL]
 
 
 HELP  = f"Gene function visualization (v{__version__})."
 
 @click.group(help=HELP)
-def cli():
+def run():
     pass
 
+def parse_genes(fname):
+    if not fname:
+        utils.stop("no input file provided.")
 
-ROOT_CHOICES = [utils.NS_BP, utils.NS_MF, utils.NS_CC, utils.NS_ALL]
+    fname = Path(fname)
 
+    if not fname.exists():
+        utils.info(f"file not found: {fname}")
 
-@cli.command()
+    stream = open(fname, "rt")
+    stream = map(lambda x: x.strip(), stream)
+    stream = filter(lambda x: x, stream)
+    stream = filter(lambda x: not x.startswith("#"), stream)
+    reader = csv.reader(stream)
+    reader = map(lambda x: x[0], reader)
+    reader = list(reader)
+
+    if len(reader) < 1:
+        utils.stop("no genes found in the input file.")
+
+    utils.debug(f"input size: {len(reader)}")
+
+    return reader
+
+@run.command()
 @click.argument("fname", default=None, required=False)
-@click.option("-o", "--out", "out", metavar="TEXT", default="genescape.pdf", help="output graph file")
-@click.option(
-    "-i",
-    "--index",
-    "index",
-    metavar="FILE",
-    help="OBO index file",
-)
-@click.option("-m", "--match", "match", metavar="REGEX", default='', help="Regular expression match on function")
-@click.option("-c", "--count", "count", metavar="INT", default=1, type=int, help="The minimal count for a GO term (1)")
-@click.option(
-    '-r',
-    '--root',
-    type=click.Choice(ROOT_CHOICES, case_sensitive=False),
-    default=utils.NS_ALL,
-    help='Select a category: BP, MF, CC, or ALL.',
-)
-@click.option("-t", "--test", "test", is_flag=True, help="run with demo data")
-@click.option("-v", "verbose", is_flag=True, help="verbose output")
-@click.help_option("-h", "--help")
-def tree(fname, out=None, index=None, root=utils.NS_ALL, match=None, count=1, verbose=False, test=False):
-    """
-    Draws a tree from GO terms.
-    """
-    # Import the tree module.
-    from genescape import tree
-
-    # Get the configuration file
-    cnf = resources.get_config()
-
-    # Initialize the resources.
-    res = resources.init(cnf)
-
-    # Get the index
-    index = resources.get_index(index=index, res=res)
-
-    # Set the verbosity level.
-    utils.verbosity(verbose)
-
-    # Override the fname if demo is set.
-    if test:
-        fname = res.TEST_GENES
-        utils.info(f"input: {fname}")
-
-    # Run the tree command.
-    graph, ann = tree.parse_input(inp=fname, index=index, pattern=match, mincount=count, root=root)
-
-    # Write the tree to a file.
-    tree.write_tree(graph, ann, out=out)
-
-
-@cli.command()
-@click.argument("fname", default=None, required=False)
-@click.option("-i", "--index", "index", metavar="TEXT", help="OBO index file.")
-@click.option("-o", "--out", "out", metavar="TEXT", default="", help="output graph file")
+@click.option("-i", "--idx", "idx_fname", metavar="TEXT", help="Genescape index file.")
[…]
