+++ pyproject.toml
@@ -38,7 +38,7 @@ Issues = "https://github.com/ialbert/genescape-central/issues"
 Source = "https://github.com/ialbert/genescape-central"
 
 [project.scripts]
-"genescape" = "genescape.main:cli"
+"genescape" = "genescape.main:run"
 
 [tool.hatch.metadata]
 include = [
