+++ src/genescape/main.py
@@ -1,13 +1,10 @@
-from genescape import __version__
-from genescape import resources, utils
-from genescape import graph
-import click, csv
 import sys, json
-from pathlib import Path
+import click
+from genescape import __version__
+from genescape import resources, utils, nexus
 
 # Valid choices for root
 ROOT_CHOICES = [utils.NS_BP, utils.NS_MF, utils.NS_CC, utils.NS_ALL]
-
 
 HELP  = f"Gene function visualization (v{__version__})."
 
@@ -15,33 +12,10 @@ HELP  = f"Gene function visualization (v{__version__})."
 def run():
     pass
 
-def parse_genes(fname):
-    if not fname:
-        utils.stop("no input file provided.")
-
-    fname = Path(fname)
-
-    if not fname.exists():
-        utils.info(f"file not found: {fname}")
-
-    stream = open(fname, "rt")
-    stream = map(lambda x: x.strip(), stream)
-    stream = filter(lambda x: x, stream)
-    stream = filter(lambda x: not x.startswith("#"), stream)
-    reader = csv.reader(stream)
-    reader = map(lambda x: x[0], reader)
-    reader = list(reader)
-
-    if len(reader) < 1:
-        utils.stop("no genes found in the input file.")
-
-    utils.debug(f"input size: {len(reader)}")
-
-    return reader
-
 @run.command()
 @click.argument("fname", default=None, required=False)
 @click.option("-i", "--idx", "idx_fname", metavar="TEXT", help="Genescape index file.")
+@click.option("-o", "--out", "out_fname", default="", metavar="TEXT", help="Output file (default: screen).")
 @click.option("-m", "--match", "match", metavar="REGEX", default='', help="Regular expression match on function")
 @click.option("-c", "--count", "count", metavar="INT", default=1, type=int, help="The minimal count for a GO term (1)")
 @click.option("-t", "--test", "test", is_flag=True, help="Run with test data")
@@ -52,7 +26,7 @@ def parse_genes(fname):
 )
 @click.option("-v", "verbose", is_flag=True, help="Verbose output.")
 @click.help_option("-h", "--help")
-def annotate(fname, idx_fname=None, root=utils.NS_ALL, verbose=False, test=False, match="", count=1):
+def annotate(fname, out_fname='', idx_fname=None, root=utils.NS_ALL, verbose=False, test=False, match="", count=1):
     """
     Generates GO terms annotations for a list of genes.
     """
@@ -63,19 +37,47 @@ def annotate(fname, idx_fname=None, root=utils.NS_ALL, verbose=False, test=False
         fname = res.TEST_GENES
         utils.info(f"input: {fname}")
 
-    targets = parse_genes(fname)
+    targets = utils.parse_genes(fname)
 
-    utils.info(f"idx: {idx_fname}")
+    pd, tree, ann = nexus.run(genes=targets, idx_fname=idx_fname, root=root,  pattern=match, mincount=count)
 
-    idx, tree = graph.load_graph(idx_fname)
+    if out_fname:
+        utils.info(f"output: {out_fname}")
+        with open(out_fname, "wt") as fp:
+            fp.write(ann)
+    else:
+        print(ann, end='')
 
-    graph.stats(idx)
+@run.command()
+@click.argument("fname", default=None, required=False)
+@click.option("-i", "--idx", "idx_fname", metavar="TEXT", help="Genescape index file.")
+@click.option("-o", "--out", "out_fname", default="output.pdf", metavar="TEXT", help="Output image file.")
[…]
