+++ src/genescape/icons.py
@@ -1,0 +1,54 @@
+"""
+An exploratory module with the goal of avoid dependency on faicons package.
+
+TODO: not fully implemented.
+"""
+import pickle
+from faicons import icon_svg
+
+# Load default icons
+icon_play = icon_svg("play")
+icon_down = icon_svg("download")
+icon_zoom_in = icon_svg("magnifying-glass-plus")
+icon_zoom_out = icon_svg("magnifying-glass-minus")
+zoom_reset = icon_svg("magnifying-glass")
+
+def pack():
+    icon_play_val=b'\x80\x04\x95C\x02\x00\x00\x00\x00\x00\x00\x8c\x0fhtmltools._core\x94\x8c\x03Tag\x94\x93\x94)\x81\x94}\x94(\x8c\x04name\x94\x8c\x03svg\x94\x8c\x06add_ws\x94\x89\x8c\x05attrs\x94h\x00\x8c\x0bTagAttrDict\x94\x93\x94)\x81\x94(\x8c\x07viewBox\x94\x8c\x0b0 0 384 512\x94\x8c\x13preserveAspectRatio\x94\x8c\x04none\x94\x8c\x0baria-hidden\x94\x8c\x04true\x94\x8c\x04role\x94\x8c\x03img\x94\x8c\x05class\x94\x8c\x02fa\x94\x8c\x05style\x94\x8c\x89fill:currentColor;height:1em;width:0.75em;margin-left:auto;margin-right:0.2em;position:relative;vertical-align:-0.125em;overflow:visible;\x94u\x8c\x08children\x94h\x00\x8c\x07TagList\x94\x93\x94)\x81\x94h\x02)\x81\x94}\x94(h\x05\x8c\x04path\x94h\x07\x88h\x08h\n)\x81\x94\x8c\x01d\x94\x8c\x92M73 39c-14.8-9.1-33.4-9.4-48.5-.9S0 62.6 0 80V432c0 17.4 9.4 33.4 24.5 41.9s33.7 8.1 48.5-.9L361 297c14.3-8.7 23-24.2 23-41s-8.7-32.2-23-41L73 39z\x94sh\x18h\x1a)\x81\x94\x8c\x10prev_displayhook\x94Nubah#Nub.'
[…]
