+++ src/genescape/nexus.py
@@ -8,6 +8,7 @@ from genescape import resources, utils
 from itertools import islice, takewhile, dropwhile, tee, chain
 import networkx as nx
 import pydot
+from pprint import pprint
 
 # Namespace categories
 NS_BP, NS_MF, NS_CC, NS_ALL = "BP", "MF", "CC", "ALL"
@@ -88,14 +89,16 @@ def build_graph(data):
     """
     # The complete ontology graph
     graph = nx.DiGraph()
-    terms = data[OBO_KEY]
+    obo = data[OBO_KEY]
 
-    terms = filter(lambda x: not x.get("is_obsolete"), terms.values())
+    # Remove obsolete terms.
+    obo = filter(lambda x: not x.get("is_obsolete"), obo.values())
+
     # nodes = islice(nodes, 100)
-    terms = list(terms)
+    obo = list(obo)
 
     # Add individual nodes to the
-    for row in terms:
+    for row in obo:
         oid = row["id"]
         name = row["name"]
         namespace = utils.NAMESPACE_MAP.get(row[NAMESPACE], "?")
@@ -108,7 +111,7 @@ def build_graph(data):
                   ANNO_TOTAL: anno_total}
         graph.add_node(oid, id=oid, name=name, namespace=namespace, **kwargs)
 
-    for row in terms:
+    for row in obo:
         child = row["id"]
         for parent in row.get("is_a", []):
             if graph.has_node(parent):
@@ -209,7 +212,11 @@ def parse_obo(stream):
             term.setdefault("is_a", []).append(target)
             continue
 
-        if code == "id" or code == "name" or code == "namespace":
+        if code == "name":
+            term[code] = " ".join(elems[1:])
+            continue
+
+        if code == "id" or code == "namespace":
             term[code] = elems[1]
             continue
 
@@ -385,7 +392,7 @@ def make_pydot(tree, status):
         src_tot = node[INP_LEN_TOT]
 
         valid_len = len(sym_valid)
-        label = f"{goid}\n{text} [{ann_count}]"
+        label = f"{goid}\n{text} [{ann_count}] [{ann_total}]"
         if is_input:
             fillcolor = utils.LEAF_COLOR if is_leaf else utils.INPUT_COLOR
             label = f"{label}\n({src_tot}/{valid_len})"
@@ -442,43 +449,7 @@ def save_json(data, path):
         fp.write(text.encode('utf-8'))
 
 
-@utils.timer
-def fill_graph(idx):
-    """
-    Fills the index with additional information
-    """
 
-    graph = build_graph(idx)
-
-    obo = idx[OBO_KEY]
-
-    # Add the degree to the nodes
-    for node_id in graph.nodes():
-        out_degree = graph.out_degree(node_id)
-        in_degree = graph.in_degree(node_id)
-        obo[node_id][OUT_DEGREE] = out_degree
-        obo[node_id][IN_DEGREE] = in_degree
-
-    # Compute the number of descendants and cumulative annotations.
-    def traverse(tree, node_id, store):
-        desc_total = 0
-        anno_total = obo[node_id][ANNO_COUNT]
-        for child in tree.successors(node_id):
-            if child not in store:
-                traverse(tree, child, store)
-            desc_total += store[child] + 1
-            anno_total += obo[child][ANNO_COUNT]
-        store[node_id] = desc_total
-        obo[node_id][DESC_COUNT] = desc_total
-        obo[node_id][ANNO_TOTAL] = anno_total
-
-    # From the roots backfill graph measures
[…]
