+++ src/genescape/main.py
@@ -8,9 +8,40 @@ from pprint import pprint
 # Valid choices for root
 ROOT_CHOICES = [utils.NS_BP, utils.NS_MF, utils.NS_CC, utils.NS_ALL]
 
-HELP  = f"Gene function visualization (v{__version__})."
+HELP = f"Gene function visualization (v{__version__})."
 
-@click.group(help=HELP)
+def check_params(it):
+    errflag = False
+    for name, fname in it:
+        if not fname:
+            utils.error(f"parameter --{name} must be specified")
+            errflag = True
+    if errflag:
+        sys.exit(1)
+
+def find_file(it):
+
+    errflag = False
+    for name, fname in it:
+        if not os.path.isfile(fname):
+            utils.error(f"file for --{name} not found: {fname}")
+            errflag = True
+
+    # Check that the files are present
+    if errflag:
+        sys.exit(1)
+
+class HelpFormatter(click.Group):
+    def format_help(self, ctx, fmt):
+        super().format_help(ctx, fmt)
+        # Here you add your additional usage examples
+        with fmt.section('Examples'):
+            fmt.write_text('genescape web')
+            fmt.write_text('genescape annotate genelist.txt')
+            fmt.write_text('genescape tree genelist.txt -o output.pdf')
+            fmt.write_text('genescape show GO:0005737')
+
+@click.group(help=HELP, cls=HelpFormatter)
 def run():
     pass
 
@@ -21,11 +52,11 @@ def run():
 @click.option("-m", "--match", "match", metavar="REGEX", default='', help="Regular expression match on function")
 @click.option("-c", "--count", "count", metavar="INT", default=1, type=int, help="The minimal count for a GO term (1)")
 @click.option("-t", "--test", "test", is_flag=True, help="Run with test data")
-@click.option( '-r', '--root',
-    type=click.Choice(ROOT_CHOICES, case_sensitive=False),
-    default=utils.NS_ALL,
-    help='Select a category: BP, MF, CC, or ALL.',
-)
+@click.option('-r', '--root',
+              type=click.Choice(ROOT_CHOICES, case_sensitive=False),
+              default=utils.NS_ALL,
+              help='Select a category: BP, MF, CC, or ALL.',
+              )
 @click.option("-v", "verbose", is_flag=True, help="Verbose output.")
 @click.help_option("-h", "--help")
 def annotate(fname, out_fname='', idx_fname=None, root=utils.NS_ALL, verbose=False, test=False, match="", count=1):
@@ -41,7 +72,7 @@ def annotate(fname, out_fname='', idx_fname=None, root=utils.NS_ALL, verbose=Fal
 
     targets = utils.parse_genes(fname)
 
-    dot, tree, ann = nexus.run(genes=targets, idx_fname=idx_fname, root=root,  pattern=match, mincount=count)
+    dot, tree, ann = nexus.run(genes=targets, idx_fname=idx_fname, root=root, pattern=match, mincount=count)
 
     if out_fname:
         utils.info(f"output: {out_fname}")
@@ -50,6 +81,7 @@ def annotate(fname, out_fname='', idx_fname=None, root=utils.NS_ALL, verbose=Fal
     else:
         print(ann, end='')
 
+
 @run.command()
 @click.argument("fname", default=None, required=False)
 @click.option("-i", "--idx", "idx_fname", metavar="TEXT", help="Genescape index file.")
@@ -57,11 +89,11 @@ def annotate(fname, out_fname='', idx_fname=None, root=utils.NS_ALL, verbose=Fal
[…]
