### 2024-05-06 56e2345
finalizing refactor
