+++ src/genescape/olap.py
@@ -1,0 +1,12 @@
+from genescape import nexus, resources
+
+def test():
+    pass
+
+
+if __name__ == '__main__':
+    res = resources.init()
+
+    fname1 = res.TEST_GENES
+
+    test()
