+++ src/genescape/shiny/tree/README.md
@@ -1,0 +1,1 @@
+Packaging this way allows us to deploy to Shinapps.io
