+++ src/genescape/shiny/README.md
@@ -1,1 +1,0 @@
-Packaging this way allows us to deploy to Shinapps.io
