+++ src/genescape/gs_index.py
@@ -1,0 +1,326 @@
+import csv, gzip, json, os
+from pathlib import Path
+from itertools import tee, takewhile, dropwhile, islice
+from genescape import utils
+import networkx as nx
+
+class Index:
+    INFO_KEY = "info"
+    OBO_KEY = "obo"
+    ANNO_COUNT = "anno_count"
+    ANNO_TOTAL = "anno_total"
+    DESC_COUNT = "desc_count"
+    OUT_DEGREE = "out_degree"
+    IN_DEGREE = "in_degree"
+    NAMESPACE = "namespace"
+
+    SYM2GO, GO2SYM, NAME2SYM = "sym2go", "go2sym", "name2sym"
+
+    def __init__(self, data=None):
+        # The data storage.
+        self.data = data or dict()
+
+        # Information on the index.
+        self.info = self.data.get(self.INFO_KEY) or dict()
+
+        # The OBO data.
+        self.obo = self.data.get(self.OBO_KEY) or dict()
+
+        # Maps symbols to GO terms.
+        self.sym2go = self.data.get(self.SYM2GO) or dict()
+
+        # Maps GO terms to symbols.
+        self.go2sym = self.data.get(self.GO2SYM) or dict()
+
+        # Maps names to symbols (synonyms).
+        self.name2sym = self.data.get(self.NAME2SYM) or dict()
+
+    def stats(self):
+        map_count = 0
+        for value in self.go2sym.values():
+            map_count += len(list(set(value)))
+        sym_count = len(self.sym2go)
+        go_count = len(self.go2sym)
+
+        return map_count, sym_count, go_count
+
+    def __str__(self):
+        map_count, sym_count, go_count = self.stats()
+        source = self.info.get("gaf_fname", "unknown")
+        return f"Index: {map_count:,d} associations of {sym_count:,d} genes over {go_count:,d} GO terms ({source})"
+
+# Parse a stream to an OBO file and for
+@utils.timer
+def parse_obo(obo_fname, idx):
+
+    # Open the file
+    stream = utils.get_stream(obo_fname)
+
+    # Remove empty rows
+    stream = filter(None, stream)
+
+    # Turn the stream into a CSV reader
+    stream = csv.reader(stream, delimiter=" ")
+
+    # Split the stream into three streams
+    head_stream, term_stream = tee(stream, 2)
+
+    # Up the reaching the first term
+    head_stream = takewhile(lambda x: x[0] != "[Term]", head_stream)
+
+    # Drop until you reach term and take until you reach the typedef
+    term_stream = dropwhile(lambda x: x[0] != "[Term]", term_stream)
+
+    # Shortcut to the info object.
+    info = idx.info
+
+    # Shortcut to the OBO object.
+    obo = idx.obo
+
+    # Read the obo file
+    info['obo_fname'] = Path(obo_fname).name
+
+    INCLUDE = ["format-version", "data-version", "ontology"]
+    INCLUDE = set(INCLUDE)
+
+    # Fill the headers
+    for elems in head_stream:
+        code = elems[0].strip(":")
+        if code in INCLUDE:
+            info[code] = elems[1]
+
+    # Read each individual OBO term.
+    term = {}
+    for elems in term_stream:
+        code = elems[0].strip(":")
+
+        # Exit this loop at the typedef stanza
+        if code == "[Typedef]":
+            break
+
+        if code == "[Term]" and term:
+            is_obsolete = term.get("is_obsolete", False)
+            if not is_obsolete:
+                obo[term["id"]] = term
+            term = {}
+            continue
+
+        if code == "is_a":
[…]
