### 2024-05-15 48a924f
just some explorations
