+++ src/genescape/gs_graph.py
@@ -1,5 +1,108 @@
 from genescape import utils, resources
 from genescape import gs_index
+import networkx as nx
+import re
+
+class NodeAttr:
+    """
+    Additional node attributes.
+    """
+    def __init__(self):
+        self.inp_flag = False
+        self.inp_sym = []
+        self.inp_len = 0
+        self.inp_sym_tot = []
+        self.inp_len_tot = 0
+        self.desc_list = []
+
+class Result:
+
+    def __init__(self, idx, targets=[], root=utils.NS_ALL, mincount=1, pattern=''):
+
+        # Collects errors
+        self.errors = []
+
+        # The index object
+        self.idx = idx
+        self.graph = idx.graph
+
+        # The input targets
+        self.targets = set(map(lambda x: x.upper(), targets))
+
+        # The missing targets
+        self.missing_targets = list(filter(lambda x: x not in self.idx.sym2go and x not in self.idx.obo, self.targets))
+
+        # Store the missing targets
+        if self.missing_targets:
+            msg = f"Missing symbols: {','.join(self.missing_targets)}"
+            self.errors.append(msg)
+
+        # The valid targets.
+        self.valid_targets = list(filter(lambda x: x in self.idx.sym2go or x in self.idx.obo, self.targets))
+
+        # Maps GO ids to input target symbols.
+        self.go2inp = dict()
+
+        # GO terms for the input targets.
+        self.goids = []
+
+        # Populate the mappings.
+        for target in self.valid_targets:
+            values = self.idx.sym2go.get(target, []) or [target]
+            for goid in values:
+                self.go2inp.setdefault(goid, []).append(target)
+                self.goids.append(goid)
+
+        # Normally the list should be empty. Sanity check.
+        self.missing_goids = list(filter(lambda x: x not in self.idx.obo, self.goids))
+
+        if self.missing_goids:
+            msg = f"Missing GO terms: {','.join(self.missing_goids)}"
+            self.errors.append(msg)
+
+        # Nodes to build the subgraph from.
+        self.valid_goids = filter(lambda x: x in self.graph.nodes, self.goids)
+
+        # Apply the root filter.
+        if root != utils.NS_ALL:
+            self.valid_goids = filter(lambda x: self.graph.nodes[x][utils.NAMESPACE] == root, self.valid_goids)
+
+        # Apply the pattern filter.
+        if pattern:
+            try:
+                patt = re.compile(pattern, re.IGNORECASE)
+                self.valid_goids = filter(lambda x: re.search(patt, self.idx.graph.nodes[x]["name"]), self.valid_goids)
+            except re.error:
+                msg = f"Invalid pattern: {pattern}"
+                self.errors.append(msg)
+
+        # Apply mincount filter.
+        self.valid_goids = filter(lambda x: len(self.go2inp[x]) >= mincount, self.valid_goids)
+
+        # Check for input.
+        if not self.valid_goids:
+            msg = "No GO terms pass all conditions."
+            self.errors.append(msg)
+
+        # Find the ancestors for each node.
+        anc = set()
+        for node in self.valid_goids:
+            anc = anc.union(nx.ancestors(self.graph, node))
+
+        # Add the original nodes basck well.
+        anc = anc.union(self.valid_goids)
+
+        # Subset the graph to the ancestors only.
+        tree = self.graph.subgraph(anc)
+
+        # Initialize the subtree
+        for node_id in tree.nodes():
[…]
