+++ src/genescape/gs_graph.py
@@ -1,19 +1,29 @@
 from genescape import utils, resources
 from genescape import gs_index
 import networkx as nx
-import re
+import re, textwrap
+import pydot
+
+ATTR_KEY = 'attr'
 
 class NodeAttr:
-    """
-    Additional node attributes.
-    """
     def __init__(self):
-        self.inp_flag = False
-        self.inp_sym = []
-        self.inp_len = 0
-        self.inp_sym_tot = []
-        self.inp_len_tot = 0
-        self.desc_list = []
+        self.is_input = False
+        self.sources = []
+        self.descendants = []
+        self.sources_all = []
+
+    @property
+    def src_len(self):
+        return len(self.sources)
+
+    @property
+    def src_all_len(self):
+        return len(self.sources_all)
+
+    @property
+    def desc_count(self):
+        return len(self.descendants)
 
 class Result:
 
@@ -53,6 +63,9 @@ class Result:
                 self.go2inp.setdefault(goid, []).append(target)
                 self.goids.append(goid)
 
+        # Remove duplicates
+        self.goids = set(self.goids)
+
         # Normally the list should be empty. Sanity check.
         self.missing_goids = list(filter(lambda x: x not in self.idx.obo, self.goids))
 
@@ -77,7 +90,7 @@ class Result:
                 self.errors.append(msg)
 
         # Apply mincount filter.
-        self.valid_goids = filter(lambda x: len(self.go2inp[x]) >= mincount, self.valid_goids)
+        self.valid_goids = set(filter(lambda x: len(self.go2inp[x]) >= mincount, self.valid_goids))
 
         # Check for input.
         if not self.valid_goids:
@@ -98,11 +111,133 @@ class Result:
         # Initialize the subtree
         for node_id in tree.nodes():
             node = tree.nodes[node_id]
-            sources = go2inp.get(node_id, [])
-            node[INP_FLAG] = node_id in valid2tgt
-            node[INP_SYM] = list(sources)
-            node[INP_LEN] = len(sources)
+            sources = self.go2inp.get(node_id, [])
+            attr = NodeAttr()
+            attr.is_input = node_id in self.valid_goids
+            attr.sources = list(sources)
+            node[ATTR_KEY] = attr
 
+        # Travers and update the cumulative totals
+        def traverse(tree, node_id, store):
+
+            store[node_id] = [node_id]
+
+            for child_id in tree.successors(node_id):
+                if child_id not in store:
+                    traverse(tree, child_id, store)
+                    store[node_id].extend(store[child_id])
+
+            # Add up the totals from the subtree.
+            totals = []
+            for parent_id in store[node_id]:
+                totals += tree.nodes[parent_id][ATTR_KEY].sources
+
+            # Shortcut to the node
+            node = tree.nodes[node_id]
+
+            # Store the descendants from the subtree only
+            node[ATTR_KEY].descendants = store[node_id]
+
+            # Add the cumulative totals
+            node[ATTR_KEY].sources_all = list(set(totals))
+
+        roots = filter(lambda x: tree.in_degree(x) == 0, tree.nodes())
+        for root in roots:
+            store = dict()
+            traverse(tree, root, store=store)
+
+        # Need a sorted tree for stability.
+        self.tree = nx.DiGraph()
+        self.tree.add_nodes_from(sorted(tree.nodes(data=True)))
[…]
