+++ src/genescape/shiny/tree/app_new.py
@@ -1,6 +1,6 @@
 from shiny import reactive
 from shiny import App, render, ui
-import asyncio, os
+import asyncio, os, time
 from genescape import icons
 from genescape import __version__, gs_graph, utils, resources
 import pandas as pd
@@ -78,7 +78,6 @@ Green nodes indicate functions present in input genes.
 
 Dark green nodes indicate leaf nodes in the ontology (highest possible granularity).
 
-
 * Subtrees are colored by Ontology namespace: Cellular Component (pink), Molecular Function (blue), Biological Process (beige).
 * The number `[234]` in a node indicates the number of annotation for that GO term in total.
 * The number `(1/5)` in a node indicates how many input genes carry that function.
@@ -120,10 +119,10 @@ app_ui = ui.page_sidebar(
         ),
 
         # The annotation data as csv.
-        ui.output_code("data_csv"),
+        ui.output_code("csv_data"),
 
         # The dot file as text.
-        ui.output_code("data_dot"),
+        ui.output_code("dot_data"),
 
         width=SIDEBAR_WIDTH, bg=SIDEBAR_BG,
     ),
@@ -155,17 +154,26 @@ app_ui = ui.page_sidebar(
                 # Error message label.
                 ui.tags.div(
                     ui.output_text(id="error_msg"),
+                    class_="error_msg",
                 ),
 
                 # Info message label.
                 ui.tags.div(
-                    ui.output_text("info_msg"),
-                    align="center"),
+                    ui.output_text(id="info_msg1"),
+                    class_="info_msg", align="center",
+                ),
 
                 # The graph root object.
-                ui.p(
+                ui.div(
                     ui.tags.b(GRAPH_TAB),
                     id="graph_root", align="center"),
+
+                # Info message label.
+                ui.tags.div(
+                    ui.output_text(id="info_msg2"),
+                    class_="info_msg", align="center",
+                ),
+                class_="graph_root",
             ),
         ),
 
@@ -181,7 +189,7 @@ app_ui = ui.page_sidebar(
         ui.nav_panel(
             "Help", ui.tags.p(
                 ui.markdown(HELP_TAB)),
-            ui.img(src="node-colors.png", ),
+
         ),
         id="tabs", selected="Graph"
     ),
@@ -197,6 +205,7 @@ app_ui = ui.page_sidebar(
     ui.head_content(
         ui.tags.script(
             """
+            // Page complete trigger.
             $(function() {
                 Shiny.addCustomMessageHandler("trigger", function(message) {
                     render_graph();
@@ -210,13 +219,159 @@ app_ui = ui.page_sidebar(
     title=PAGE_TITLE, id="main",
 )
 
+def text2list(text):
+    """
+    Parses a text inptu into a list of terms
+    """
+    terms = map(lambda x: x.strip(), text.splitlines())
+    terms = filter(None, terms)
+    terms = list(terms)
+    return terms
 
 def server(input, output, session):
-    pass
+    # The annotation table.
+    ann_df = reactive.Value(pd.DataFrame())
 
+    # Runtime messages
+    info_value1 = reactive.Value("")
+
+    # Runtime messages
+    info_value2 = reactive.Value("")
+
+    # Runtime error message
+    err_value = reactive.Value("")
+
+    # Annotation as a CSV string (invisible)
+    csv_value = reactive.Value("# The annotation CSV will appear here.")
+
+    # The dot file as a text (invisible)
+    dot_value = reactive.Value("# The dot file will appear here.")
+
+    @reactive.effect
+    @reactive.event(input.submit)
+    async def submit():
+        await create_tree()
+
+    async def create_tree():
+        limit = 10
[…]
