+++ src/genescape/shiny/tree/app_new.py
@@ -49,30 +49,38 @@ Press "Draw Tree" to generate the graph.
 
 ANNOT_TAB = """
 
-## Functional annotations
+Genes may be annotated with Gene Ontology functions. 
 
-This panel shows the functional annotations for the input genes.
+This panel shows the cumulative functional annotations across the list.
 
-Each gene is annotated with GO terms. The coverage indicates how many genes in the input cover that function.
+The coverage column indicates how many genes in the input cover that function.
 
-You can filter the table by entering a regular expression in the pattern field.
+How to reduce the graph size:
+
+1. The **Coverage** filter will select for the minimum coverage.
+1. The **Filtering pattern** will keep only functions that match.
+
+Examples: 
+
+* `Coverage=5` at least 5 genes carrying the function
+* `Filter=GTP|kinase` keep only functions that match both `GTP` and `kinase`.
+
 """
 
 HELP_TAB = """
-
-## How to use GeneScape
+Type a list of genes of GO terms then press "Draw Tree".
 
 Tips for making graphs smaller:
 
 1. Filter for minimum coverage in the functions. You can use regular expressions in the pattern.
 
-## Legend
+**Legend**
 
 <img src="/static/node-help.png" class="img-fluid helpimg" alt="Alt text">
 
 The coverage indicates how many genes in the input cover that function.
 
-## Colors
+**Colors**
 
 Green nodes indicate functions present in input genes.
 
@@ -99,14 +107,26 @@ app_ui = ui.page_sidebar(
         # Select the database.
         ui.input_select(
             "database",
-            "Organism", DATABASE_CHOICES,
+            label="Organism:", choices=DATABASE_CHOICES,
+        ),
+
+        # Select the root
+        ui.input_select(
+            "root",
+            "Domain:", {
+                utils.NS_ALL: "All domains",
+                utils.NS_BP: "Biological Process",
+                utils.NS_MF: "Molecular Function",
+                utils.NS_CC: "Cellular Component"
+            },
+            selected=utils.NS_BP,
         ),
 
         # Minimum count.
-        ui.input_text("coverage", label="Coverage", value=MINCOUNT),
+        ui.input_text("coverage", label="Coverage:", value=MINCOUNT),
 
         # Filtering pattern.
-        ui.input_text("pattern", label="Pattern (regex ok)", value=PATTERN),
+        ui.input_text("pattern", label="Word filter (regex ok):", value=PATTERN),
 
         # Download button for annotation CSV.
         ui.tags.p(
@@ -133,22 +153,22 @@ app_ui = ui.page_sidebar(
         ui.nav_panel(
             "Graph", ui.tags.p(
 
-                ui.tags.p(
+                ui.tags.div(
                     ui.input_action_button("zoom_in", label="Zoom", icon=icons.icon_zoom_in,
-                                           class_="btn btn-light",
+                                           class_="btn btn-light btn-sm",
                                            data_action="zoom_in"),
                     ui.tags.span(" "),
-                    ui.input_action_button("zoom_reset", label="Reset", icon=icons.zoom_reset,
-                                           class_="btn btn-light",
+                    ui.input_action_button("zoom_reset", label=" Reset", icon=icons.zoom_reset,
+                                           class_="btn btn-light btn-sm",
                                            data_action="zoom_reset"),
                     ui.tags.span(" "),
                     ui.input_action_button("zoom_out", label="Zoom", icon=icons.icon_zoom_out,
-                                           class_="btn btn-light ",
+                                           class_="btn btn-light btn-sm",
[…]
