+++ src/genescape/shiny/tree/app-old.py
@@ -1,0 +1,312 @@
+from shiny import reactive
+from shiny import App, render, ui
+import asyncio, os
+from genescape import icons
+from genescape import __version__, nexus, utils, resources
+import pandas as pd
+
+# Get a resource configuration file.
+cnf_fname = os.environ.get("GENESCAPE_CONFIG", None)
+
+# Load the configuration file.
+cnf = resources.get_config(fname=cnf_fname)
+
+# Load the default resources.
+res = resources.init(config=cnf)
+
+# Update the resources from the environment.
+res.update_from_env()
+
+# This is the global database choices
+DATABASE_CHOICES = dict()
+
+# Update the choices with the defaults.
+DATABASE_CHOICES.update(res.index_choices())
+
+PAGE_TITLE = res.config.get("PAGE_TITLE", "GeneScape")
+
+# Default gene list.
+GENE_LIST = res.config.get("GENE_LIST", "ABTB3\nBCAS4\nC3P1\nGRTP1")
+
+# Default pattern.
+PATTERN = res.config.get("PATTERN", "")
+
+# Default mincount.
+MINCOUNT = res.config.get("MINCOUNT", 1)
+
+# Load the sidebar width.
+SIDEBAR_WIDTH = res.config.get("SIDEBAR_WIDTH", 300)
+
+# Default sidebar background color.
+SIDEBAR_BG = res.config.get("SIDEBAR_WIDTH", "#f9f9f9")
+
+# Home page link
+HOME = "https://github.com/ialbert/genescape-central/"
+
+HELP = """
+The functional annotations will be shown in the table below. 
+"""
+
+DOCS = """
+* Reduce the graph size by filtering for coverage or words in the functions (regex ok).
+* The coverage indicates how many genes in the input cover that function.
+* Green nodes indicate functions present in input genes.
+* Dark green nodes indicate leaf nodes in the ontology (highest possible granularity).
+* Subtrees are colored by Ontology namespace: Cellular Component (pink), Molecular Function (blue), Biological Process (beige).
+* The number `[234]` in a node indicates the number of annotation for that GO term in total.
+* The number `(1/5)` in a node indicates how many input genes carry that function.
+"""
+
+ANN_COLUMNS = "coverage,function,node_id,source,size,ann_count,ann_total,desc_count".split(",")
+
+app_ui = ui.page_sidebar(
+
+    ui.sidebar(
+        ui.input_text_area("terms", label="Gene List", value=GENE_LIST),
+
+        ui.div({"class": "left-aligned"},
+
+               ui.input_select(
+                   "database",
+                   "Organism", DATABASE_CHOICES,
+               ),
+
+               ui.input_text("mincount", label="Coverage", value=MINCOUNT),
+               ui.input_text("pattern", label="Pattern", value=PATTERN),
+
+               ui.input_select(
+                   "root",
+                   "Root", {
+                       utils.NS_ALL: "All roots",
+                       utils.NS_BP: "Biological Process",
+                       utils.NS_MF: "Molecular Function",
+                       utils.NS_CC: "Cellular Component"
+                   },
+               ),
+               ),
+        ui.input_action_button("submit", "Draw Tree", class_="btn-success", icon=icons.icon_play),
+
+        ui.tags.p(
+            ui.download_link("download_csv", "Download annotations as CSV", icon=icons.icon_down),
+        ),
+        ui.tags.p(
+            ui.download_link("download_dot", "Download graph as a DOT file", icon=icons.icon_down),
+        ),
+        ui.output_code("annot_csv"),
+        ui.output_code("dot_elem"),
+
+        width=SIDEBAR_WIDTH, bg=SIDEBAR_BG,
+    ),
+
+    ui.head_content(
+        ui.tags.script(
[…]
