+++ src/genescape/shiny/tree/app_new.py
@@ -1,402 +1,0 @@
-from shiny import reactive
-from shiny import App, render, ui
-import asyncio, os, time
-from genescape import icons
-from genescape import __version__, gs_graph, utils, resources
-import pandas as pd
-from pathlib import Path
-
-from starlette.applications import Starlette
-from starlette.routing import Mount
-from starlette.staticfiles import StaticFiles
-
-# Load the default resources.
-res = resources.init()
-
-# This is the global database choices
-DATABASE_CHOICES = dict()
-
-# Update the choices with the defaults.
-DATABASE_CHOICES.update(res.index_choices())
-
-PAGE_TITLE = res.config.get("PAGE_TITLE", "GeneScape")
-
-# Default gene list.
-GENE_LIST = res.config.get("GENE_LIST", "ABTB3\nBCAS4\nC3P1\nGRTP1")
-
-# Default pattern.
-PATTERN = res.config.get("PATTERN", "")
-
-# Default mincount.
-MINCOUNT = res.config.get("MINCOUNT", "autodetect")
-
-# Load the sidebar width.
-SIDEBAR_WIDTH = res.config.get("SIDEBAR_WIDTH", 300)
-
-# Default sidebar background color.
-SIDEBAR_BG = res.config.get("SIDEBAR_WIDTH", "#f9f9f9")
-
-# Home page link
-HOME = "https://github.com/ialbert/genescape-central/"
-
-HELP = """
-The functional annotations will be shown in the table below. 
-"""
-
-GRAPH_TAB = """
-Press "Draw Tree" to generate the graph.
-"""
-
-ANNOT_TAB = """
-
-Genes may be annotated with Gene Ontology functions. 
-
-This panel shows the cumulative functional annotations across the list.
-
-The coverage column indicates how many genes in the input cover that function.
-
-How to reduce the graph size:
-
-1. The **Coverage** filter will select for the minimum coverage.
-1. The **Filtering pattern** will keep only functions that match.
-
-Examples: 
-
-* `Coverage=5` at least 5 genes carrying the function
-* `Filter=GTP|kinase` keep only functions that match both `GTP` and `kinase`.
-
-"""
-
-HELP_TAB = """
-Type a list of genes of GO terms then press "Draw Tree".
-
-Tips for making graphs smaller:
-
-1. Filter for minimum coverage in the functions. You can use regular expressions in the pattern.
-
-**Legend**
-
-<img src="/static/node-help.png" class="img-fluid helpimg" alt="Alt text">
-
-The coverage indicates how many genes in the input cover that function.
-
-**Colors**
-
-Green nodes indicate functions present in input genes.
-
-Dark green nodes indicate leaf nodes in the ontology (highest possible granularity).
-
-* Subtrees are colored by Ontology namespace: Cellular Component (pink), Molecular Function (blue), Biological Process (beige).
-* The number `[234]` in a node indicates the number of annotation for that GO term in total.
-* The number `(1/5)` in a node indicates how many input genes carry that function.
-"""
-
-print(os.getcwd())
-
-app_ui = ui.page_sidebar(
-    ui.sidebar(
-
-        # The gene list.
-        ui.input_text_area("input_list", label="Gene List", value=GENE_LIST),
-
-        # The main submit button.
-        ui.input_action_button(
-            "submit", "Draw Tree", class_="btn-success", icon=icons.icon_play
-        ),
-
-        # Select the database.
-        ui.input_select(
-            "database",
-            label="Organism:", choices=DATABASE_CHOICES,
-        ),
-
-        # Select the root
-        ui.input_select(
-            "root",
-            "Domain:", {
-                utils.NS_ALL: "All domains",
-                utils.NS_BP: "Biological Process",
[…]
