+++ README.md
@@ -14,7 +14,7 @@ The public version of the software can be accessed at:
 
 [pyshiny]: https://shiny.posit.co/py/
 
-## Local installation
+### Local installation
 
 Users can also run the program on their system by installing the software via `pip`:
 
@@ -48,7 +48,7 @@ The user interface is browser-based via the URL:
 
 Once you are done running with the program, you will need to close the terminal that started the program or press CTRL+C to stop the program from running.
 
-## What does GeneScape do?
+### What does GeneScape do?
 
 **GeneScape** works the following way:
 
@@ -58,9 +58,9 @@ Once you are done running with the program, you will need to close the terminal 
 
 > **Note**: Even short lists of genes (under ten genes) can create large trees. Filter by minimum counts (how many genes share the function) or functional patterns (functions that match a pattern). 
 
-When not explicitly specified, GeneScape will try to find reasonable coverage treshold for the input genes.
+When not explicitly specified, GeneScape will try to find reasonable coverage threshold for the input genes.
 
-## Node Labeling 
+### Node Labeling 
 
 The labels in the graph carry additional information on the number of genes in the input that carry that function as well as an indicator for the specificity of the function in the organism. For example, the label:
 
@@ -77,7 +77,7 @@ The `(1/5)` means that  `1` out of `5` genes in the input list carry this annota
 
 It is possible to compute a p-value to determine whether an observed enrichment difference is statistically significant. We'll just note that assigning p-values to enrichment counts is fraught with several challenges. In our opinion, GO annotations are neither complete, nor independent, nor precise enough to satisfy mathematical requirements. In addition, appropriate selection of the background (aka the `19000` above) to correct p-values for multiple comparisons also presents many challenges. For these reasons, we do not compute p-values in our application.
 
-## Node Coloring
+### Node Coloring
 
 The colors in the tree carry additional meaning:
 
@@ -113,7 +113,7 @@ Filters are applied during the annotation step and will filter the GO terms deri
 
 In the Shiny interface use the **coverage** filter to remove functions that are not well represented in the input list. Recall that `coverage` represents the number of genes in the input list that carry that function. You can see the counts for each annotation in the **Function Annotations** box as the first column.
 
-## Command line requirements
+### Command line requirements
 
 To generate images from command line the `graphviz` software must be installed. You can install it via `conda`
 
@@ -133,7 +133,7 @@ Then use an online tool like [viz-js][viz] to visualize the graph.
 
 [viz]: https://viz-js.com/ 
 
-## genescape tree
+### genescape tree
 
 We packaged test data with the software so you can test it like so:
 
@@ -147,7 +147,7 @@ Which will generate a tree visualization of the test data.
 
 [out1]: https://raw.githubusercontent.com/ialbert/genescape-central/main/docs/images/gs_output_1.png
 
-## Reducing the graph size
+### Reducing the graph size
 
 We can pass the tree visualizer a list of genes or a list of GO ids, or even a mix of both. 
 
@@ -166,17 +166,16 @@ For most gene list the result might be a huge tree
[…]
