+++ src/genescape/__about__.py
@@ -1,1 +1,1 @@
-__version__ = "0.9.5"
+__version__ = "0.9.8"
