+++ docs/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'GeneScape: A Python package for gene ontology analysis'
+title: 'GeneScape: A Python package for gene ontology visualization'
 tags:
   - Python
   - biology
@@ -35,7 +35,7 @@ The [Gene Ontology Consortium][GO] maintains GAF files for various organisms. Ty
 
 # Statement of need
 
-The most annotated gene in the human genome, `HTT1`, currently has 1098 annotations. Thus, even small lists of genes may have a large number of annotations presenting an extraordinary challenge for interpretation. There is a clear need to visualize shared gene functions in an informative manner. 
+The most annotated gene in the human genome, `HTT`, currently has 1100 annotations. Thus, even small lists of genes may have a large number of annotations presenting an extraordinary challenge for interpretation. There is a clear need to visualize shared gene functions in an informative manner. 
 
 Web-based tools designed to visualize and filter gene ontology data include `AmiGO` [@AmiGO] and `QuickGO` [@QuickGO]. Command line tools like `goatools` [@goatools] support GO term lineage visualization. R packages like `topGO` [@topGO] implement GO structure visualizations of enriched GO terms. We are unaware of locally installable software that specifically allows for interactive filtering and visualization of gene ontology derived on gene lists.
 
@@ -72,28 +72,43 @@ GRTP1
 Users can process the list above via the command line or via the Shiny interface. A command line invocation might look like:
 
 ```console
-genescape tree genes.txt -o output.pdf
+genescape tree genes1.txt -o output.pdf
 ```
+
+The command above will produce the image:
+
+![Ontology subgraph for a gene list \label{fig:interface}][img_shiny]{height="216pt"}
+
+[img_shiny]: images/gs_output_1.png
 
 Internally, GeneScape first transforms the input gene list into a GO term list, where additional information is added to each term:
 
 ```
-count,function,root,goid,source,size,label
-1,activation of GTPase activity,BP,GO:0090630,GRTP1,4,(1/4)
-1,protein heterodimerization activity,MF,GO:0046982,ABTB3,4,(1/4)
-1,BLOC-1 complex,CC,GO:0031083,BCAS4,4,(1/4)
-1,membrane,CC,GO:0016020,ABTB3,4,(1/4)
-1,cytoplasm,CC,GO:0005737,BCAS4,4,(1/4)
-1,extracellular space,CC,GO:0005615,C3P1,4,(1/4)
-1,GTPase activator activity,MF,GO:0005096,GRTP1,4,(1/4)
-1,endopeptidase inhibitor activity,MF,GO:0004866,C3P1,4,(1/4)
+Coverage,Function,Domain,GO,Genes
+1,endopeptidase inhibitor activity,MF,GO:0004866,C3P1
+1,GTPase activator activity,MF,GO:0005096,GRTP1
+1,extracellular space,CC,GO:0005615,C3P1
+1,cytoplasm,CC,GO:0005737,BCAS4
+1,membrane,CC,GO:0016020,ABTB3
+1,PDZ domain binding,MF,GO:0030165,ABTB3
+1,BLOC-1 complex,CC,GO:0031083,BCAS4
+1,"synaptic transmission, glutamatergic",BP,GO:0035249,ABTB3
+1,exploration behavior,BP,GO:0035640,ABTB3
+1,protein heterodimerization activity,MF,GO:0046982,ABTB3
+1,protein stabilization,BP,GO:0050821,ABTB3
+1,activation of GTPase activity,BP,GO:0090630,GRTP1
+1,glutamatergic synapse,CC,GO:0098978,ABTB3
 ```
 
[…]
