+++ Makefile
@@ -40,9 +40,12 @@ tag: test
 
 # Generate images for the documentation
 docimg:
-	genescape tree -o docs/images/gs_output_1.png src/genescape/data/test_genes.txt
-	genescape tree -o docs/images/gs_output_2.png src/genescape/data/test_genes.txt --mincov 1
-	genescape tree -m repair --mincov 2 -o docs/images/gs_output_3.png src/genescape/data/test_genes.txt
+	genescape tree -o docs/images/gs_output_1.png docs/genelist_1.txt
+	genescape tree -o docs/images/gs_output_2.png docs/genelist_2.txt --mincov 1
+	genescape tree -m lipid --mincov 2 -o docs/images/gs_output_3.png docs/genelist_2.txt
+
+pdf:
+	(cd docs && pandoc paper.md -o paper.pdf)
 
 # Runs a linter.
 lint:
