+++ src/genescape/shiny/tree/static/README.txt
@@ -1,0 +1,1 @@
+# Placeholder file.
