### 2024-05-22 #6624 @ialbert
@mahfuz05062  Checklist completed.

Software release: 1.0.4

* PyPI: https://pypi.org/project/genescape/

Zenodo DOI for the software: `10.5281/zenodo.11245263`

* Zenodo: https://zenodo.org/doi/10.5281/zenodo.11245263

Thank you reviewers @j-andrews7 @sridhar0605 for the suggestions. The software is now so much more usable.
