+++ docs/paper.bib
@@ -9,7 +9,6 @@
 	Date-Modified = {2024-03-28 18:29:13 +0000},
 	Doi = {10.1038/75556},
 	Id = {Ashburner2000},
-	Isbn = {1546-1718},
 	Journal = {Nature Genetics},
 	Number = {1},
 	Pages = {25--29},
