### 2024-06-01 #6624 @ialbert
@mahfuz05062 I believe I located the source of the error, the paper.bib has additional fields, and it seems that the ISBN number  does not fully conform.  I have now removed the ISBN field.

Try to see whether the acceptance now works, and I'll keep troubleshooting the bibtex if not.
