### 2024-06-05 #6624 @ialbert
> as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. I have checked this review, your repository, the archive link, and the paper. Most seems in order, however the below requires your attention:
>
> * [x ]  Please edit the ZENODO archive listed version tag to read `v1.0.4` (i.e. to include the `v`).

@Kevin-Mattheus-Moerman version edited as requested

https://zenodo.org/records/11245264
