### 2024-06-05 ca32873
added citation to readme
