### 2024-06-05 #6624 @ialbert
Thanks again to all participants;

This was my first time publishing in JOSS, thanks to direct author/reviewer interactions, the outcomes were both speedier and better than that of a typical scientific software publication process.

Here is hoping that the methods championed via JOSS become the norm.
